refactor(orchestrator): split lifecycle concerns
