+++ internal/orchestrator/orchestrator.go
@@ -3,20 +3,15 @@ package orchestrator
 import (
 	"context"
 	"errors"
-	"fmt"
 	"log/slog"
-	"slices"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	workflowconfig "github.com/digitaldrywood/detent/internal/config"
 	"github.com/digitaldrywood/detent/internal/connector"
 	"github.com/digitaldrywood/detent/internal/gate"
 	runpkg "github.com/digitaldrywood/detent/internal/runner"
-	"github.com/digitaldrywood/detent/internal/runtimeoutput"
 	"github.com/digitaldrywood/detent/internal/scheduler"
 	"github.com/digitaldrywood/detent/internal/selector"
 	"github.com/digitaldrywood/detent/internal/store"
@@ -182,70 +177,6 @@ type configUpdateRequest struct {
 type runUpdate struct {
 	issueID string
 	usage   runpkg.UsageUpdate
-}
-
-func ConfigFromWorkflow(cfg workflowconfig.Config) Config {
-	identity := cfg.Identity
-	identity.Normalize()
-
-	return Config{
-		PollInterval:               durationFromMillis(cfg.Polling.IntervalMS),
-		MaxConcurrentAgents:        cfg.Agent.MaxConcurrentAgents,
-		MaxConcurrentAgentsByState: cloneStateLimits(cfg.Agent.MaxConcurrentAgentsByState),
-		DispatchPriorityByState:    append([]string(nil), cfg.Agent.DispatchPriorityByState...),
-		DispatchPriorityByLabel:    append([]string(nil), cfg.Agent.DispatchPriorityByLabel...),
-		MergeFastPathEnabled:       cfg.Agent.MergeFastPath.Enabled,
-		MaxConcurrentAgentsPerHost: positiveIntValue(cfg.Worker.MaxConcurrentAgentsPerHost),
-		MaxRetryBackoff:            durationFromMillis(cfg.Agent.MaxRetryBackoffMS),
-		Claiming: ClaimingConfig{
-			Enabled:           cfg.Tracker.Claims.Enabled,
-			OwnershipMode:     identity.OwnershipMode,
-			Owner:             identity.Name,
-			AssigneeLogin:     identity.GitHubLogin,
-			OwnerField:        identity.OwnerField,
-			LeaseField:        cfg.Tracker.Claims.LeaseField,
-			LeaseTTL:          durationFromSeconds(cfg.Tracker.Claims.TTLSeconds),
-			HeartbeatInterval: durationFromSeconds(cfg.Tracker.Claims.HeartbeatSeconds),
-		},
-		AutoPromote: normalizeAutoPromoteConfig(AutoPromoteConfig{
-			Enabled:            cfg.Agent.AutoPromote.Enabled,
-			QuietDuration:      durationFromSeconds(cfg.Agent.AutoPromote.QuietSeconds),
-			OptoutLabel:        cfg.Agent.AutoPromote.OptoutLabel,
-			AllowedIssueLabels: append([]string(nil), cfg.Agent.AutoPromote.AllowedIssueLabels...),
-			SourceState:        cfg.Agent.AutoPromote.SourceState,
-			PassState:          cfg.Agent.AutoPromote.PassState,
-			ReworkState:        cfg.Agent.AutoPromote.ReworkState,
-			ReworkLimit:        cfg.Agent.AutoPromote.ReworkLimit,
-			Gate:               gate.Effective(cfg.Gate),
-		}),
-		Plan: gate.EffectivePlan(cfg.Plan),
-		DependencyAutoUnblock: normalizeDependencyAutoUnblockConfig(DependencyAutoUnblockConfig{
-			Enabled:      cfg.Tracker.DependencyAutoUnblock.Enabled,
-			SourceStates: append([]string(nil), cfg.Tracker.DependencyAutoUnblock.SourceStates...),
-			TargetState:  cfg.Tracker.DependencyAutoUnblock.TargetState,
-			Readiness:    cfg.Tracker.DependencyAutoUnblock.Readiness,
-		}),
-		BlockerAutoPromote: BlockerAutoPromoteConfig{
-			Enabled:       cfg.Tracker.BlockerAutoPromote.Enabled,
-			SourceStates:  append([]string(nil), cfg.Tracker.BlockerAutoPromote.SourceStates...),
-			BlockerStates: append([]string(nil), cfg.Tracker.BlockerAutoPromote.BlockerStates...),
-			TargetState:   cfg.Tracker.BlockerAutoPromote.TargetState,
-		},
-		ActiveStates:                  append([]string(nil), cfg.Tracker.ActiveStates...),
-		ObservedStates:                append([]string(nil), cfg.Tracker.ObservedStates...),
-		TerminalStates:                append([]string(nil), cfg.Tracker.TerminalStates...),
-		Authorization:                 cfg.Tracker.Authorization,
-		SelectorContext:               selector.Context{InstanceLogin: identity.GitHubLogin, Persona: identity.Name},
-		WorkerHosts:                   append([]string(nil), cfg.Worker.SSHHosts...),
-		BudgetRefusalCooldown:         durationFromSeconds(cfg.Budget.RefusalCooldownSeconds),
-		WorkspaceCleanupIdleTTL:       durationFromMillis(cfg.Workspace.CleanupIdleTTLMS),
-		WorkspaceCleanupSweepInterval: durationFromMillis(cfg.Workspace.CleanupSweepIntervalMS),
-		SelectorPersona:               cfg.Tracker.Assignee,
-		GitHubGraphQLWarnRemaining:    int64(cfg.Tracker.GitHubGraphQLWarnRemaining),
-		GitHubGraphQLMinReserve:       int64(cfg.Tracker.GitHubGraphQLMinReserve),
-		GitHubRESTMinReserve:          int64(cfg.Tracker.GitHubRESTMinReserve),
-		OutputTruncationMaxBytes:      cfg.Agent.OutputTruncation.MaxBytes,
-	}
 }
 
 func New(cfg Config, deps Dependencies) (*Orchestrator, error) {
@@ -481,831 +412,6 @@ func (o *Orchestrator) ForceQuit(ctx context.Context) error {
 	}
 }
 
-func (o *Orchestrator) observedStatusFetchStates() []string {
-	states := append([]string{blockedStatusState}, autoPromoteFetchStates(o.cfg.AutoPromote)...)
-	if cfg := gate.EffectivePlan(o.cfg.Plan); cfg.Enabled {
-		states = append(states, cfg.Stop)
-	}
-	states = append(states, o.cfg.ObservedStates...)
-	cfg := normalizeDependencyAutoUnblockConfig(o.cfg.DependencyAutoUnblock)
-	if cfg.Enabled {
-		states = append(states, cfg.SourceStates...)
-	}
-	blockerCfg := normalizeBlockerAutoPromoteConfig(o.cfg.BlockerAutoPromote, o.cfg.ActiveStates, o.cfg.DependencyAutoUnblock)
-	if blockerCfg.Enabled {
-		states = append(states, blockerCfg.SourceStates...)
-		states = append(states, blockerCfg.BlockerStates...)
-	}
-	return displayStateNames(states)
-}
-
-func (o *Orchestrator) observedStatusFetchStatesForTick(state *State) []string {
-	states := o.observedStatusFetchStates()
-	if !autoPromoteUsesMergePassState(o.cfg.AutoPromote) || o.mergeWorkerLocalSlotsAvailable(state) {
-		return states
-	}
-	return statesWithoutState(states, autoPromoteMergingState)
-}
-
-func (o *Orchestrator) mergeWorkerLocalSlotsAvailable(state *State) bool {
-	stats := o.projectStateSlotStats(connector.Issue{State: autoPromoteMergingState}, state)
-	return stats.available > 0
-}
-
-func statesWithoutState(states []string, omit string) []string {
-	omit = normalizeState(omit)
-	out := make([]string, 0, len(states))
-	for _, state := range states {
-		if normalizeState(state) == omit {
-			continue
-		}
-		out = append(out, state)
-	}
-	return out
-}
-
-func autoPromoteUsesMergePassState(cfg AutoPromoteConfig) bool {
-	cfg = normalizeAutoPromoteConfig(cfg)
-	return normalizeState(cfg.PassState) == normalizeState(autoPromoteMergingState)
-}
-
-func autoPromoteFetchStates(cfg AutoPromoteConfig) []string {
-	cfg = normalizeAutoPromoteConfig(cfg)
-	candidates := []string{cfg.SourceState}
-	if autoPromoteUsesMergePassState(cfg) {
-		candidates = append(candidates, cfg.PassState)
-	}
-	states := make([]string, 0, len(candidates))
-	seen := make(map[string]struct{}, len(candidates))
-	for _, state := range candidates {
-		key := normalizeState(state)
-		if key == "" {
-			continue
-		}
-		if _, ok := seen[key]; ok {
-			continue
-		}
-		seen[key] = struct{}{}
-		states = append(states, state)
-	}
-	return states
-}
-
-func displayStateNames(states []string) []string {
-	out := make([]string, 0, len(states))
-	seen := make(map[string]struct{}, len(states))
-	for _, state := range states {
-		key := normalizeState(state)
-		if key == "" {
-			continue
-		}
-		if _, ok := seen[key]; ok {
-			continue
-		}
-		seen[key] = struct{}{}
-		out = append(out, displayStateName(state))
-	}
-	return out
-}
-
-func displayStateName(state string) string {
-	state = strings.TrimSpace(state)
-	switch normalizeState(state) {
-	case "blocked":
-		return blockedStatusState
-	case "human review":
-		return autoPromoteSourceState
-	case "merging":
-		return autoPromoteMergingState
-	case "rework":
-		return autoPromoteReworkState
-	case "todo":
-		return "Todo"
-	case "in progress":
-		return "In Progress"
-	default:
-		return state
-	}
-}
-
-func issuesInStates(issues []connector.Issue, states []string) []connector.Issue {
-	wanted := stateNameSet(states)
-	if len(wanted) == 0 {
-		return nil
-	}
-
-	out := make([]connector.Issue, 0, len(issues))
-	for _, issue := range issues {
-		if _, ok := wanted[normalizeState(issue.State)]; ok {
-			out = append(out, cloneIssue(issue))
-		}
-	}
-	return out
-}
-
-func blockedStatusTransitionIssues(blocked map[string]Blocked) []connector.Issue {
-	out := make([]connector.Issue, 0, len(blocked))
-	for _, entry := range blocked {
-		legacyStatusIssue := entry.Source == "" && normalizeState(entry.Issue.State) == normalizeState(blockedStatusState)
-		if entry.Source != BlockedSourceProjectStatus && !legacyStatusIssue {
-			continue
-		}
-		if strings.TrimSpace(entry.Issue.ID) == "" {
-			continue
-		}
-		out = append(out, cloneIssue(entry.Issue))
-	}
-	return out
-}
-
-func stateNameSet(states []string) map[string]struct{} {
-	out := make(map[string]struct{}, len(states))
-	for _, state := range states {
-		state = normalizeState(state)
-		if state != "" {
-			out[state] = struct{}{}
-		}
-	}
-	return out
-}
-
-func (o *Orchestrator) reconcileClosedCompletedIssueStatuses(ctx context.Context, state *State, issues []connector.Issue, now time.Time) map[string]struct{} {
-	targetState := doneStateName(o.cfg.TerminalStates)
-	reconciled := map[string]struct{}{}
-	for _, issue := range issues {
-		issueID := strings.TrimSpace(issue.ID)
-		if issueID == "" {
-			continue
-		}
-		if _, ok := reconciled[issueID]; ok {
-			continue
-		}
-		if !closedCompletedIssueNeedsStatusReconciliation(issue, o.cfg.TerminalStates) {
-			continue
-		}
-		if err := o.updateIssueStateByID(ctx, issueID, issue, targetState, now, "closed_completed_status_reconciled"); err != nil {
-			if o.logger != nil {
-				o.logger.Warn("reconcile closed completed issue status failed", "issue_id", issueID, "identifier", issue.Identifier, "from_state", issue.State, "target_state", targetState, "error", err)
-			}
-			continue
-		}
-		reconciled[issueID] = struct{}{}
-		if o.logger != nil {
-			o.logger.Info("reconciled closed completed issue status", "issue_id", issueID, "identifier", issue.Identifier, "from_state", issue.State, "target_state", targetState)
-		}
-		recordStateEvent(state, telemetry.ActivityEvent{
-			At:      now,
-			Event:   "closed_completed_status_reconciled",
-			Message: "reconciled " + issueLabel(issue) + " from " + strings.TrimSpace(issue.State) + " to " + targetState,
-		})
-	}
-	if len(reconciled) == 0 {
-		return nil
-	}
-	return reconciled
-}
-
-func closedCompletedIssueNeedsStatusReconciliation(issue connector.Issue, terminalStates []string) bool {
-	return issue.Closed &&
-		closedReasonCompleted(issue.ClosedReason) &&
-		strings.TrimSpace(issue.State) != "" &&
-		!stateIn(issue.State, terminalStates)
-}
-
-func closedReasonCompleted(reason string) bool {
-	reason = strings.ToLower(strings.TrimSpace(reason))
-	reason = strings.ReplaceAll(reason, "-", "_")
-	return reason == "completed"
-}
-
-func filterReconciledIssues(issues []connector.Issue, reconciled map[string]struct{}) []connector.Issue {
-	if len(reconciled) == 0 || len(issues) == 0 {
-		return issues
-	}
-	out := issues[:0]
-	for _, issue := range issues {
-		if _, ok := reconciled[strings.TrimSpace(issue.ID)]; ok {
-			continue
-		}
-		out = append(out, issue)
-	}
-	return out
-}
-
-func recordStateEvent(state *State, event telemetry.ActivityEvent) {
-	if state == nil {
-		return
-	}
-	state.RecentEvents = append(state.RecentEvents, event)
-	if len(state.RecentEvents) > maxRecentEvents {
-		state.RecentEvents = append([]telemetry.ActivityEvent(nil), state.RecentEvents[len(state.RecentEvents)-maxRecentEvents:]...)
-	}
-}
-
-func issueLabel(issue connector.Issue) string {
-	if identifier := strings.TrimSpace(issue.Identifier); identifier != "" {
-		return identifier
-	}
-	return strings.TrimSpace(issue.ID)
-}
-
-func (o *Orchestrator) reconcileRunningIssues(ctx context.Context, state *State, now time.Time) {
-	ids := runningIssueIDs(state.Running)
-	if len(ids) == 0 {
-		return
-	}
-	if !o.shouldReconcileRunningIssues(state, now) {
-		return
-	}
-	state.LastRunningReconcileAt = now
-
-	issues, err := o.connector.FetchIssueStatesByIDs(ctx, ids)
-	if err != nil {
-		if o.logger != nil {
-			o.logger.Warn("fetch running issue states failed", "error", err)
-		}
-		return
-	}
-
-	byID := make(map[string]connector.Issue, len(issues))
-	for _, issue := range issues {
-		if strings.TrimSpace(issue.ID) == "" {
-			continue
-		}
-		byID[issue.ID] = issue
-	}
-
-	for _, id := range ids {
-		issue, ok := byID[id]
-		if !ok {
-			continue
-		}
-
-		running := state.Running[id]
-		running.Issue = o.hydrateRunningIssueComments(ctx, mergeIssueTrackerFields(running.Issue, issue))
-		if workspaceIssueTerminal(running.Issue, o.cfg.TerminalStates) {
-			o.completeTerminalRunning(ctx, state, id, running, terminalCompletedAt(running.Issue, o.cfg.TerminalStates, now), running.Tokens)
-			continue
-		}
-		state.Running[id] = running
-
-		if claimed, ok := state.Claimed[id]; ok {
-			claimed.Issue = mergeIssueTrackerFields(claimed.Issue, issue)
-			state.Claimed[id] = claimed
-		}
-	}
-}
-
-func (o *Orchestrator) shouldReconcileRunningIssues(state *State, now time.Time) bool {
-	if len(state.Running) == 0 {
-		return false
-	}
-	if state.LastRunningReconcileAt.IsZero() {
-		return true
-	}
-	interval := max(o.cfg.PollInterval, defaultRunningReconcileInterval)
-	return !now.Before(state.LastRunningReconcileAt.Add(interval))
-}
-
-func (o *Orchestrator) hydrateRunningIssueComments(ctx context.Context, issue connector.Issue) connector.Issue {
-	if len(issue.Comments) > 0 {
-		return issue
-	}
-	reader, ok := o.connector.(connector.IssueCommentReader)
-	if !ok {
-		return issue
-	}
-	comments, err := reader.FetchIssueComments(ctx, issue)
-	if err != nil {
-		if o.logger != nil {
-			o.logger.Warn("fetch running issue comments failed", "issue_id", issue.ID, "identifier", issue.Identifier, "error", err)
-		}
-		return issue
-	}
-	issue.Comments = comments
-	return issue
-}
-
-func mergeIssueTrackerFields(current, refreshed connector.Issue) connector.Issue {
-	merged := cloneIssue(current)
-	refreshed = cloneIssue(refreshed)
-
-	if strings.TrimSpace(refreshed.ID) != "" {
-		merged.ID = refreshed.ID
-	}
-	if refreshed.Identifier != "" {
-		merged.Identifier = refreshed.Identifier
-	}
-	if refreshed.Title != "" {
-		merged.Title = refreshed.Title
-	}
-	if refreshed.Description != "" {
-		merged.Description = refreshed.Description
-	}
-	if refreshed.Priority != nil {
-		merged.Priority = refreshed.Priority
-	}
-	if refreshed.State != "" {
-		merged.State = refreshed.State
-	}
-	if refreshed.BranchName != "" {
-		merged.BranchName = refreshed.BranchName
-	}
-	if refreshed.URL != "" {
-		merged.URL = refreshed.URL
-	}
-	merged.Closed = refreshed.Closed
-	if refreshed.ClosedReason != "" {
-		merged.ClosedReason = refreshed.ClosedReason
-	}
-	if refreshed.PRNumber != nil {
-		merged.PRNumber = refreshed.PRNumber
-	}
-	if refreshed.PullRequest != nil {
-		merged.PullRequest = refreshed.PullRequest
-	}
-	if refreshed.AuthorID != "" {
-		merged.AuthorID = refreshed.AuthorID
-	}
-	if refreshed.AssigneeID != "" {
-		merged.AssigneeID = refreshed.AssigneeID
-	}
-	if refreshed.Assignees != nil {
-		merged.Assignees = refreshed.Assignees
-	}
-	if refreshed.BlockedBy != nil {
-		merged.BlockedBy = refreshed.BlockedBy
-	}
-	if refreshed.BlockerReason != "" {
-		merged.BlockerReason = refreshed.BlockerReason
-	}
-	if refreshed.Labels != nil {
-		merged.Labels = refreshed.Labels
-	}
-	if refreshed.Fields != nil {
-		merged.Fields = refreshed.Fields
-	}
-	if refreshed.CreatedAt != nil {
-		merged.CreatedAt = refreshed.CreatedAt
-	}
-	if refreshed.UpdatedAt != nil {
-		merged.UpdatedAt = refreshed.UpdatedAt
-	}
-	if refreshed.StageUpdatedAt != nil {
-		merged.StageUpdatedAt = refreshed.StageUpdatedAt
-	}
-	if refreshed.ModelOverride != "" {
-		merged.ModelOverride = refreshed.ModelOverride
-	}
-
-	return merged
-}
-
-func runningIssueIDs(running map[string]Running) []string {
-	ids := sortedKeys(running)
-	out := ids[:0]
-	for _, id := range ids {
-		if strings.TrimSpace(id) != "" {
-			out = append(out, id)
-		}
-	}
-	return out
-}
-
-func (o *Orchestrator) markRefresh(state *State, now time.Time) {
-	state.PollInterval = o.cfg.PollInterval
-	state.MaxConcurrentAgents = o.cfg.MaxConcurrentAgents
-	if o.cfg.PollInterval > 0 {
-		state.NextRefreshAt = now.Add(o.cfg.PollInterval)
-		return
-	}
-	state.NextRefreshAt = time.Time{}
-}
-
-func (o *Orchestrator) markRefreshSucceeded(state *State, now time.Time) {
-	state.DataSeq++
-	state.LastRefreshAt = now.UTC()
-}
-
-func (o *Orchestrator) finishRefresh(state *State, now time.Time) {
-	o.captureConnectorAuthHealth(state)
-	cycle := o.captureConnectorRateLimits(state, now)
-	o.logGraphQLRateLimitCycle(cycle)
-	restCycle := o.captureConnectorRESTRateLimits(state, now)
-	o.logRESTRateLimitCycle(restCycle)
-
-	interval := o.adaptivePollInterval(state, now)
-	state.PollInterval = interval
-	if interval > 0 {
-		state.NextRefreshAt = now.Add(interval)
-		return
-	}
-	state.NextRefreshAt = time.Time{}
-}
-
-func (o *Orchestrator) captureConnectorAuthHealth(state *State) {
-	if reporter, ok := o.connector.(connector.AuthHealthReporter); ok {
-		if health, ok := reporter.AuthHealth(); ok {
-			state.Auth = health
-			return
-		}
-	}
-	state.Auth = connector.AuthHealth{}
-}
-
-type graphQLRateLimitCycle struct {
-	Bucket     *telemetry.RateLimitBucket
-	Cost       *telemetry.GraphQLCost
-	HasSummary bool
-}
-
-type restRateLimitCycle struct {
-	Bucket     *telemetry.RateLimitBucket
-	Usage      *telemetry.RESTUsage
-	HasSummary bool
-}
-
-func (o *Orchestrator) captureConnectorRateLimits(state *State, now time.Time) graphQLRateLimitCycle {
-	var usage connector.GraphQLRateLimitUsage
-	if reporter, ok := o.connector.(connector.GraphQLRateLimitUsageReporter); ok {
-		usage = reporter.FlushGraphQLRateLimitUsage()
-	}
-
-	var rateLimit connector.GraphQLRateLimit
-	hasRateLimit := usage.HasRateLimit
-	status := graphQLRateLimitTelemetryStatus(usage.RateLimitStatus)
-	if hasRateLimit {
-		rateLimit = usage.RateLimit
-	} else if status == "" {
-		reporter, ok := o.connector.(connector.RateLimitReporter)
-		if !ok {
-			return graphQLRateLimitCycle{}
-		} else {
-			var okRateLimit bool
-			rateLimit, okRateLimit = reporter.GraphQLRateLimit()
-			if !okRateLimit {
-				return graphQLRateLimitCycle{}
-			}
-			hasRateLimit = okRateLimit
-		}
-	}
-
-	cost := graphQLCostSummary(usage)
-	bucket := gitHubGraphQLBucket(rateLimit, now, status)
-	if cost != nil {
-		bucket.Cost = cost.TotalCost
-	}
-	if state.RateLimits == nil {
-		state.RateLimits = &telemetry.RateLimits{}
-	}
-	state.RateLimits.GitHubGraphQL = bucket
-	state.RateLimits.GraphQLCost = cost
-	return graphQLRateLimitCycle{
-		Bucket:     bucket,
-		Cost:       cost,
-		HasSummary: hasRateLimit,
-	}
-}
-
-func (o *Orchestrator) captureConnectorRESTRateLimits(state *State, now time.Time) restRateLimitCycle {
-	reporter, ok := o.connector.(connector.RESTRateLimitUsageReporter)
-	if !ok {
-		return restRateLimitCycle{}
-	}
-	usage := reporter.FlushRESTRateLimitUsage()
-	if !usage.HasRateLimit && usage.TotalRequests == 0 && len(usage.Requests) == 0 {
-		return restRateLimitCycle{}
-	}
-
-	bucket := gitHubRESTBucket(usage, now)
-	summary := restUsageSummary(usage)
-	if state.RateLimits == nil {
-		state.RateLimits = &telemetry.RateLimits{}
-	}
-	state.RateLimits.GitHubREST = bucket
-	state.RateLimits.RESTUsage = summary
-	return restRateLimitCycle{
-		Bucket:     bucket,
-		Usage:      summary,
-		HasSummary: true,
-	}
-}
-
-func restUsageSummary(usage connector.RESTRateLimitUsage) *telemetry.RESTUsage {
-	if usage.TotalRequests == 0 && len(usage.Requests) == 0 && !usage.RateLimited {
-		return nil
-	}
-
-	out := &telemetry.RESTUsage{
-		TotalRequests: usage.TotalRequests,
-		RateLimited:   usage.RateLimited,
-		Contributors:  make([]telemetry.RESTUsageContributor, 0, len(usage.Requests)),
-	}
-	if !usage.BackoffUntil.IsZero() {
-		backoffUntil := usage.BackoffUntil
-		out.BackoffUntil = &backoffUntil
-	}
-	for _, request := range usage.Requests {
-		contributor := telemetry.RESTUsageContributor{
-			EndpointFamily: request.EndpointFamily,
-			Count:          request.Count,
-			Remaining:      request.Remaining,
-			Limit:          request.Limit,
-			Resource:       request.Resource,
-			RateLimited:    request.RateLimited,
-			LastStatus:     request.LastStatus,
-		}
-		if !request.ResetAt.IsZero() {
-			resetAt := request.ResetAt
-			contributor.ResetAt = &resetAt
-		}
-		if request.RetryAfter > 0 {
-			contributor.RetryAfterMS = request.RetryAfter.Milliseconds()
-		}
-		out.Contributors = append(out.Contributors, contributor)
-	}
-	return out
-}
-
-func gitHubRESTBucket(usage connector.RESTRateLimitUsage, now time.Time) *telemetry.RateLimitBucket {
-	rateLimit := usage.RateLimit
-	var resetAt *time.Time
-	var observedAt *time.Time
-	var resetInSeconds int64
-	if !rateLimit.ResetAt.IsZero() {
-		value := rateLimit.ResetAt
-		resetAt = &value
-	}
-	if !rateLimit.UpdatedAt.IsZero() {
-		value := rateLimit.UpdatedAt
-		observedAt = &value
-	}
-	if rateLimit.RetryAfter > 0 {
-		updatedAt := rateLimit.UpdatedAt
-		if updatedAt.IsZero() {
-			updatedAt = now
-		}
-		value := updatedAt.Add(rateLimit.RetryAfter)
-		resetAt = &value
-		resetInSeconds = int64(rateLimit.RetryAfter.Round(time.Second) / time.Second)
-	}
-	if !usage.BackoffUntil.IsZero() && usage.BackoffUntil.After(now) {
-		value := usage.BackoffUntil
-		resetAt = &value
-		resetInSeconds = int64(usage.BackoffUntil.Sub(now).Round(time.Second) / time.Second)
-	}
-
-	return &telemetry.RateLimitBucket{
-		Remaining:      rateLimit.Remaining,
-		Used:           rateLimit.Used,
-		Limit:          rateLimit.Limit,
-		Cost:           usage.TotalRequests,
-		ResetAt:        resetAt,
-		ObservedAt:     observedAt,
-		ResetInSeconds: resetInSeconds,
-	}
-}
-
-func (o *Orchestrator) logRESTRateLimitCycle(cycle restRateLimitCycle) {
-	if !cycle.HasSummary || cycle.Bucket == nil {
-		return
-	}
-
-	var resetAt time.Time
-	if cycle.Bucket.ResetAt != nil {
-		resetAt = *cycle.Bucket.ResetAt
-	}
-	totalRequests := int64(0)
-	rateLimited := false
-	var contributors []telemetry.RESTUsageContributor
-	if cycle.Usage != nil {
-		totalRequests = cycle.Usage.TotalRequests
-		rateLimited = cycle.Usage.RateLimited
-		contributors = cycle.Usage.Contributors
-	}
-
-	o.logger.Info(
-		"github rest budget summary",
-		"request_count", totalRequests,
-		"rate_limited", rateLimited,
-		"remaining", cycle.Bucket.Remaining,
-		"limit", cycle.Bucket.Limit,
-		"reset_at", resetAt,
-		"contributors", contributors,
-	)
-}
-
-func graphQLCostSummary(usage connector.GraphQLRateLimitUsage) *telemetry.GraphQLCost {
-	if usage.TotalQueries == 0 && usage.TotalCost == 0 && len(usage.QueryCosts) == 0 {
-		return nil
-	}
-
-	cost := &telemetry.GraphQLCost{
-		TotalQueries: usage.TotalQueries,
-		TotalCost:    usage.TotalCost,
-		Contributors: make([]telemetry.GraphQLCostContributor, 0, len(usage.QueryCosts)),
-	}
-	for _, contributor := range usage.QueryCosts {
-		cost.Contributors = append(cost.Contributors, telemetry.GraphQLCostContributor{
-			QueryType: contributor.QueryType,
-			Count:     contributor.Count,
-			Cost:      contributor.Cost,
-		})
-	}
-	return cost
-}
-
-func (o *Orchestrator) logGraphQLRateLimitCycle(cycle graphQLRateLimitCycle) {
-	if !cycle.HasSummary || cycle.Bucket == nil {
-		return
-	}
-
-	var resetAt time.Time
-	if cycle.Bucket.ResetAt != nil {
-		resetAt = *cycle.Bucket.ResetAt
-	}
-	cycleCost := int64(0)
-	queryCount := int64(0)
-	var contributors []telemetry.GraphQLCostContributor
-	if cycle.Cost != nil {
-		cycleCost = cycle.Cost.TotalCost
-		queryCount = cycle.Cost.TotalQueries
-		contributors = cycle.Cost.Contributors
-	}
-
-	o.logger.Debug(
-		"github graphql budget summary",
-		"cycle_cost", cycleCost,
-		"query_count", queryCount,
-		"remaining", cycle.Bucket.Remaining,
-		"limit", cycle.Bucket.Limit,
-		"reset_at", resetAt,
-		"contributors", contributors,
-	)
-
-	if cycle.Bucket.Remaining < o.cfg.GitHubGraphQLWarnRemaining {
-		o.logger.Warn(
-			"github graphql budget below warning floor",
-			"remaining", cycle.Bucket.Remaining,
-			"warning_floor", o.cfg.GitHubGraphQLWarnRemaining,
-			"limit", cycle.Bucket.Limit,
-			"reset_at", resetAt,
-		)
-	}
-}
-
-func graphQLRateLimitTelemetryStatus(status string) string {
-	switch strings.TrimSpace(status) {
-	case connector.GraphQLRateLimitStatusUnknown:
-		return telemetry.RateLimitStatusUnknown
-	case connector.GraphQLRateLimitStatusExhausted:
-		return telemetry.RateLimitStatusExhausted
-	case connector.GraphQLRateLimitStatusBackoff:
-		return telemetry.RateLimitStatusBackoff
-	default:
-		return ""
-	}
-}
-
-func gitHubGraphQLBucket(rateLimit connector.GraphQLRateLimit, now time.Time, status string) *telemetry.RateLimitBucket {
-	var resetAt *time.Time
-	var observedAt *time.Time
-	var resetInSeconds int64
-	if !rateLimit.ResetAt.IsZero() {
-		value := rateLimit.ResetAt
-		resetAt = &value
-	}
-	if !rateLimit.UpdatedAt.IsZero() {
-		value := rateLimit.UpdatedAt
-		observedAt = &value
-	}
-	if rateLimit.RetryAfter > 0 {
-		updatedAt := rateLimit.UpdatedAt
-		if updatedAt.IsZero() {
-			updatedAt = now
-		}
-		value := updatedAt.Add(rateLimit.RetryAfter)
-		resetAt = &value
-		resetInSeconds = int64(rateLimit.RetryAfter.Round(time.Second) / time.Second)
-	}
-	if status == "" {
-		status = graphQLRateLimitStatusFromSnapshot(rateLimit)
-	}
-
-	return &telemetry.RateLimitBucket{
-		Remaining:      rateLimit.Remaining,
-		Used:           rateLimit.Used,
-		Limit:          rateLimit.Limit,
-		Cost:           rateLimit.Cost,
-		Status:         status,
-		ResetAt:        resetAt,
-		ObservedAt:     observedAt,
-		ResetInSeconds: resetInSeconds,
-	}
-}
-
-func graphQLRateLimitStatusFromSnapshot(rateLimit connector.GraphQLRateLimit) string {
-	if rateLimit.Limit > 0 && rateLimit.Remaining <= 0 {
-		return telemetry.RateLimitStatusExhausted
-	}
-	if rateLimit.RetryAfter > 0 {
-		return telemetry.RateLimitStatusBackoff
-	}
-	return ""
-}
-
-func (o *Orchestrator) adaptivePollInterval(state *State, now time.Time) time.Duration {
-	base := o.cfg.PollInterval
-	if base <= 0 {
-		base = defaultPollInterval
-	}
-
-	if pause := o.gitHubGraphQLPause(state, now); pause > base {
-		return pause
-	}
-	if pause := o.gitHubRESTPause(state, now); pause > base {
-		return pause
-	}
-
-	bucket := gitHubGraphQLBucketFromState(state)
-	if bucket == nil || bucket.Remaining <= 0 || bucket.Remaining >= gitHubGraphQLBackoffRemaining {
-		return base
-	}
-
-	multiplier := int64(gitHubGraphQLBackoffRemaining) / bucket.Remaining
-	if int64(gitHubGraphQLBackoffRemaining)%bucket.Remaining != 0 {
-		multiplier++
-	}
-	if multiplier < 2 {
-		multiplier = 2
-	}
-	return base * time.Duration(multiplier)
-}
-
-func (o *Orchestrator) gitHubRESTPause(state *State, now time.Time) time.Duration {
-	bucket := gitHubRESTBucketFromState(state)
-	if bucket == nil || bucket.ResetAt == nil {
-		return 0
-	}
-	if bucket.ResetInSeconds > 0 && bucket.ResetAt.After(now) {
-		return bucket.ResetAt.Sub(now)
-	}
-	if bucket.Remaining > 0 {
-		return 0
-	}
-	if !bucket.ResetAt.After(now) {
-		return 0
-	}
-	return bucket.ResetAt.Sub(now)
-}
-
-func gitHubRESTBucketFromState(state *State) *telemetry.RateLimitBucket {
-	if state.RateLimits == nil {
-		return nil
-	}
-	return state.RateLimits.GitHubREST
-}
-
-func gitHubRESTRemaining(state *State) int64 {
-	bucket := gitHubRESTBucketFromState(state)
-	if bucket == nil {
-		return 0
-	}
-	return bucket.Remaining
-}
-
-func (o *Orchestrator) gitHubGraphQLPause(state *State, now time.Time) time.Duration {
-	bucket := gitHubGraphQLBucketFromState(state)
-	if bucket == nil || bucket.ResetAt == nil {
-		return 0
-	}
-	if bucket.ResetInSeconds > 0 && bucket.ResetAt.After(now) {
-		return bucket.ResetAt.Sub(now)
-	}
-	if bucket.Remaining >= gitHubGraphQLPauseRemaining {
-		return 0
-	}
-	if !bucket.ResetAt.After(now) {
-		return 0
-	}
-	return bucket.ResetAt.Sub(now)
-}
-
-func gitHubGraphQLBucketFromState(state *State) *telemetry.RateLimitBucket {
-	if state.RateLimits == nil {
-		return nil
-	}
-	return state.RateLimits.GitHubGraphQL
-}
-
-func gitHubGraphQLRemaining(state *State) int64 {
-	bucket := gitHubGraphQLBucketFromState(state)
-	if bucket == nil {
-		return 0
-	}
-	return bucket.Remaining
-}
-
 func (o *Orchestrator) applyRuntimeUpdate(state *State, update RuntimeUpdate, ticker *time.Ticker) {
 	cfg := normalizeConfig(update.Config)
 	o.cfg = cfg
@@ -1329,87 +435,6 @@ func (o *Orchestrator) applyRuntimeUpdate(state *State, update RuntimeUpdate, ti
 		state.NextRefreshAt = state.LastRefreshAt.Add(cfg.PollInterval)
 	}
 	ticker.Reset(cfg.PollInterval)
-}
-
-func (o *Orchestrator) trackBlockedCandidates(state *State, issues []connector.Issue, now time.Time) {
-	o.dispatchPlanner().trackBlockedCandidates(state, issues, now)
-}
-
-func (o *Orchestrator) trackBlockedStatusIssues(state *State, issues []connector.Issue, now time.Time) {
-	seenBlocked := make(map[string]struct{}, len(issues))
-	for _, issue := range issues {
-		if issue.ID == "" {
-			continue
-		}
-		seenBlocked[issue.ID] = struct{}{}
-		o.setBlockedStatusIssue(state, issue, now)
-	}
-
-	for issueID, blocked := range state.Blocked {
-		if blocked.Source != BlockedSourceProjectStatus {
-			continue
-		}
-		if _, ok := seenBlocked[issueID]; !ok {
-			delete(state.Blocked, issueID)
-		}
-	}
-}
-
-func (o *Orchestrator) trackCandidateBlockedStatusIssues(state *State, issues []connector.Issue, now time.Time) {
-	blockedState := normalizeState(blockedStatusState)
-	for _, issue := range issues {
-		if issue.ID == "" {
-			continue
-		}
-		issueState := normalizeState(issue.State)
-		if issueState == "" {
-			continue
-		}
-		if issueState != blockedState {
-			clearBlockedStatusIssue(state, issue.ID)
-			continue
-		}
-		o.setBlockedStatusIssue(state, issue, now)
-	}
-}
-
-func clearBlockedStatusIssue(state *State, issueID string) {
-	if blocked, ok := state.Blocked[issueID]; ok && blocked.Source == BlockedSourceProjectStatus {
-		delete(state.Blocked, issueID)
-	}
-}
-
-func (o *Orchestrator) setBlockedStatusIssue(state *State, issue connector.Issue, now time.Time) {
-	if existing, ok := state.Blocked[issue.ID]; ok &&
-		existing.Source == BlockedSourceProjectStatus &&
-		strings.HasPrefix(existing.Reason, instantFailureBlockedReasonPrefix) &&
-		strings.TrimSpace(issue.BlockerReason) == "" {
-		existing.Issue = cloneIssue(issue)
-		state.Blocked[issue.ID] = existing
-		return
-	}
-	recovery := EvaluateBlockedRecovery(issue)
-	state.Blocked[issue.ID] = Blocked{
-		Issue:          cloneIssue(issue),
-		Reason:         blockedStatusReason(issue),
-		RecoveryReason: string(recovery.Reason),
-		RecoveryTarget: recovery.TargetState,
-		BlockedAt:      now,
-		Source:         BlockedSourceProjectStatus,
-	}
-}
-
-func blockedFromDependency(blocked Blocked) bool {
-	return blocked.Source == BlockedSourceDependency ||
-		(blocked.Source == "" && blocked.Reason == blockedReasonDependency)
-}
-
-func blockedStatusReason(issue connector.Issue) string {
-	reason := strings.TrimSpace(issue.BlockerReason)
-	if reason != "" {
-		return reason
-	}
-	return blockedReasonProjectStatus
 }
 
 func (o *Orchestrator) startDrain(state *State, now time.Time) {
@@ -1481,854 +506,6 @@ func (o *Orchestrator) abandonClaim(ctx context.Context, issueID string) error {
 	return nil
 }
 
-func (o *Orchestrator) dispatchReadyIssues(ctx context.Context, state *State, issues []connector.Issue, now time.Time) {
-	o.markGlobalProjectIdle()
-	if state.Draining {
-		return
-	}
-	planner := o.dispatchPlanner()
-	var lastDispatchFailure string
-	planner.plan(state, issues, now, dispatchPlanHooks{
-		hydrate: func(issue connector.Issue) (connector.Issue, bool) {
-			return o.hydrateDispatchIssue(ctx, issue)
-		},
-		beforeDispatch: func(_ connector.Issue, continuationIndex int) bool {
-			if continuationIndex < 0 {
-				return true
-			}
-			return waitForDispatchBackoff(ctx, continuationDelay(continuationIndex))
-		},
-		dispatch: func(issue connector.Issue, attempt int, workerHost string) bool {
-			outcome := o.dispatchIssueWithOutcome(ctx, state, issue, attempt, now, workerHost)
-			if !outcome.dispatched {
-				lastDispatchFailure = outcome.reason
-			} else {
-				lastDispatchFailure = ""
-			}
-			return outcome.dispatched
-		},
-		dispatchFailed: func(issue connector.Issue) bool {
-			return !mergeWorkerIssue(issue) || lastDispatchFailure != dispatchIssueFailureGlobalSlotUnavailable
-		},
-		retryDispatchFailed: func(issue connector.Issue, retry Retry) {
-			planner.scheduleRetry(state, issue, retry.Attempt, now, "claim verification failed", false, retry.WorkerHost)
-		},
-		preserveMissingDueRetry: func(retry Retry) bool {
-			return o.preserveMissingDueRetry(state, retry)
-		},
-		decision: func(decision dispatchPlanDecision) {
-			o.logDispatchPlanDecision(ctx, state, now, decision)
-		},
-	})
-}
-
-func (o *Orchestrator) preserveMissingDueRetry(state *State, retry Retry) bool {
-	if normalizeState(retry.Issue.State) != normalizeState(autoPromoteMergingState) {
-		return false
-	}
-	return !o.mergeWorkerLocalSlotsAvailable(state)
-}
-
-func (o *Orchestrator) hydrateDispatchIssue(ctx context.Context, issue connector.Issue) (connector.Issue, bool) {
-	if strings.TrimSpace(issue.ID) == "" || len(issue.Fields) > 0 || o.connector == nil {
-		return issue, true
-	}
-	issues, err := o.connector.FetchIssueStatesByIDs(ctx, []string{issue.ID})
-	if err != nil {
-		if o.logger != nil {
-			o.logger.Warn("hydrate dispatch issue failed", "issue_id", issue.ID, "error", err)
-		}
-		return connector.Issue{}, false
-	}
-	for _, hydrated := range issues {
-		if hydrated.ID == issue.ID {
-			return mergeIssueTrackerFields(issue, hydrated), true
-		}
-	}
-	return connector.Issue{}, false
-}
-
-func (o *Orchestrator) dispatchCandidates(ctx context.Context, state *State, issues []connector.Issue, now time.Time) {
-	o.markGlobalProjectIdle()
-	if state.Draining {
-		return
-	}
-	for _, issue := range issues {
-		if availableSlots(state) == 0 {
-			return
-		}
-		issue, ok := o.hydrateDispatchIssue(ctx, issue)
-		if !ok {
-			continue
-		}
-		if !o.dispatchable(issue, state, now) {
-			continue
-		}
-
-		o.dispatchIssue(ctx, state, issue, 0, now, "")
-	}
-}
-
-func dueRetriesByIssue(state *State, now time.Time) map[string]Retry {
-	retries := make(map[string]Retry, len(state.Retry))
-	for _, retry := range state.Retry {
-		if !retry.DueAt.After(now) {
-			retries[retry.Issue.ID] = retry
-		}
-	}
-	return retries
-}
-
-func (o *Orchestrator) dispatchable(issue connector.Issue, state *State, now time.Time) bool {
-	return o.dispatchPlanner().dispatchable(issue, state, now)
-}
-
-const (
-	dispatchIssueFailureLocalSlotUnavailable  = "local_slot_unavailable"
-	dispatchIssueFailureWorkerHostUnavailable = "worker_host_unavailable"
-	dispatchIssueFailureGlobalSlotUnavailable = "global_slot_unavailable"
-	dispatchIssueFailureClaimFailed           = "claim_failed"
-	dispatchIssueFailureWorkAttemptStart      = "work_attempt_start_failed"
-	dispatchIssueFailureStartStateTransition  = "start_state_transition_failed"
-)
-
-type dispatchIssueOutcome struct {
-	dispatched bool
-	reason     string
-}
-
-func (o *Orchestrator) dispatchIssue(
-	ctx context.Context,
-	state *State,
-	issue connector.Issue,
-	attempt int,
-	now time.Time,
-	preferredWorkerHost string,
-) bool {
-	return o.dispatchIssueWithOutcome(ctx, state, issue, attempt, now, preferredWorkerHost).dispatched
-}
-
-func (o *Orchestrator) dispatchIssueWithOutcome(
-	ctx context.Context,
-	state *State,
-	issue connector.Issue,
-	attempt int,
-	now time.Time,
-	preferredWorkerHost string,
-) dispatchIssueOutcome {
-	runMode := o.dispatchMode(state, issue)
-	targetState := dispatchStartTransitionState(issue, runMode, o.cfg.ActiveStates)
-	slotIssue := issue
-	if targetState != "" {
-		slotIssue = cloneIssue(issue)
-		slotIssue.State = targetState
-		if !o.dispatchPlanner().slotsAvailable(slotIssue, state, preferredWorkerHost) {
-			return dispatchIssueOutcome{reason: dispatchIssueFailureLocalSlotUnavailable}
-		}
-	}
-	projectStats := o.projectStateSlotStats(slotIssue, state)
-
-	workerHost, ok := o.selectWorkerHost(state, preferredWorkerHost)
-	if !ok {
-		o.logMergeWorkerFailure(issue, "worker_host_unavailable", nil)
-		o.recordMergeFailed(state, issue, now, "worker_host_unavailable", nil)
-		return dispatchIssueOutcome{reason: dispatchIssueFailureWorkerHostUnavailable}
-	}
-
-	globalSlot, ok, decision := o.acquireGlobalDispatchSlot(ctx, slotIssue, workerHost, now)
-	if !ok {
-		o.logSchedulerSlotDecision(issue, "waiting", decision, projectStats)
-		if mergeWorkerIssue(issue) {
-			o.logMergeWorkerSlotWait(issue, decision, projectStats)
-		} else {
-			o.logDispatchSlotWait(issue, decision, projectStats)
-			o.recordDispatchSlotWait(state, issue, decision, projectStats, now)
-		}
-		return dispatchIssueOutcome{reason: dispatchIssueFailureGlobalSlotUnavailable}
-	}
-	mergeTiming := o.markMergeWorkerSlotAcquired(state, issue, now)
-	o.logSchedulerSlotDecision(issue, "acquired", decision, projectStats)
-	o.logMergeWorkerSlotAcquired(issue, decision, projectStats, mergeTiming)
-	o.logWorkerLifecycle(issue, "worker_slot_acquired",
-		"attempt", attempt,
-		"worker_host", strings.TrimSpace(workerHost),
-	)
-
-	claimedIssue, claim, ok := o.claimIssue(ctx, issue, now)
-	if !ok {
-		o.releaseGlobalDispatchSlot(globalSlot)
-		o.logWorkerLifecycle(issue, "worker_capacity_released",
-			"attempt", attempt,
-			"worker_host", strings.TrimSpace(workerHost),
-			"reason", dispatchIssueFailureClaimFailed,
-		)
-		o.logMergeWorkerFailure(issue, "claim_failed", nil)
-		o.recordMergeFailed(state, issue, now, "claim_failed", nil)
-		return dispatchIssueOutcome{reason: dispatchIssueFailureClaimFailed}
-	}
-
-	issue = cloneIssue(claimedIssue)
-	workAttemptID, ok := o.startDurableWorkAttempt(ctx, state, issue, attempt, now, workerHost, runMode)
-	if !ok {
-		o.releaseGlobalDispatchSlot(globalSlot)
-		o.logWorkerLifecycle(issue, "worker_capacity_released",
-			"attempt", attempt,
-			"worker_host", strings.TrimSpace(workerHost),
-			"reason", dispatchIssueFailureWorkAttemptStart,
-		)
-		if abandonErr := o.abandonClaim(ctx, issue.ID); abandonErr != nil && o.logger != nil {
-			o.logger.Warn("abandon claim after work attempt start failed", "issue_id", issue.ID, "error", abandonErr)
-		}
-		o.logMergeWorkerFailure(issue, "work_attempt_start_failed", nil)
-		o.recordMergeFailed(state, issue, now, "work_attempt_start_failed", nil)
-		return dispatchIssueOutcome{reason: dispatchIssueFailureWorkAttemptStart}
-	}
-	if targetState != "" {
-		if err := o.updateIssueState(ctx, issue, targetState, now, "dispatch_start"); err != nil {
-			o.releaseGlobalDispatchSlot(globalSlot)
-			o.completeDurableWorkAttempt(ctx, state, Running{
-				Issue:         issue,
-				Attempt:       attempt,
-				WorkAttemptID: workAttemptID,
-				Mode:          runMode,
-				StartedAt:     now,
-				WorkerHost:    workerHost,
-			}, now, store.WorkAttemptTerminalFailure, workAttemptErrorStartTransition, err.Error(), "starting", "start state transition failed")
-			o.logWorkerLifecycle(issue, "worker_capacity_released",
-				"attempt", attempt,
-				"worker_host", strings.TrimSpace(workerHost),
-				"reason", dispatchIssueFailureStartStateTransition,
-			)
-			if abandonErr := o.abandonClaim(ctx, issue.ID); abandonErr != nil && o.logger != nil {
-				o.logger.Warn("abandon claim after start state transition failed", "issue_id", issue.ID, "error", abandonErr)
-			}
-			if o.logger != nil {
-				o.logger.Warn("start state transition failed", "issue_id", issue.ID, "identifier", issue.Identifier, "from_state", issue.State, "target_state", targetState, "error", err)
-			}
-			o.logMergeWorkerFailure(issue, "start_state_transition_failed", err)
-			o.recordMergeFailed(state, issue, now, "start_state_transition_failed", err)
-			return dispatchIssueOutcome{reason: dispatchIssueFailureStartStateTransition}
-		}
-		issue.State = targetState
-	}
-	o.markMergeStarted(state, issue, now)
-	claim.Issue = issue
-	runCtx, cancel := context.WithCancel(ctx)
-	state.Running[issue.ID] = Running{
-		Issue:         issue,
-		Attempt:       attempt,
-		WorkAttemptID: workAttemptID,
-		Mode:          runMode,
-		StartedAt:     now,
-		WorkerHost:    workerHost,
-		globalSlot:    globalSlot,
-		cancel:        cancel,
-	}
-	o.setGlobalDispatchPreempt(globalSlot, cancel)
-	state.Claimed[issue.ID] = claim
-	delete(state.Retry, issue.ID)
-	delete(state.Blocked, issue.ID)
-	delete(state.BudgetRefusals, issue.ID)
-	delete(state.ReapedWorkspaces, issue.ID)
-	delete(state.Completed, issue.ID)
-
-	request := RunRequest{
-		Issue:           issue,
-		Attempt:         attempt,
-		WorkAttemptID:   workAttemptID,
-		Mode:            runMode,
-		PriorAttempt:    state.PriorAttempts[issue.ID],
-		StartedAt:       now,
-		WorkerHost:      workerHost,
-		SelectorContext: o.selectorContext(),
-		OnUsageUpdate:   o.usageUpdateHandler(runCtx, issue.ID),
-	}
-	o.logMergeWorkerAttempt(issue, attempt, workerHost)
-	o.logWorkerLifecycle(issue, "worker_attempt_started",
-		"attempt", attempt,
-		"worker_host", strings.TrimSpace(workerHost),
-		"mode", strings.TrimSpace(runMode),
-	)
-	o.supervisor.Dispatch(runCtx, request, o.runResults)
-	return dispatchIssueOutcome{dispatched: true}
-}
-
-func dispatchStartTransitionState(issue connector.Issue, mode string, activeStates []string) string {
-	if mode != runpkg.RunModeImplement {
-		return ""
-	}
-	if normalizeState(issue.State) != "todo" {
-		return ""
-	}
-	if !stateIn(planImplementationState, activeStates) {
-		return ""
-	}
-	return planImplementationState
-}
-
-func (o *Orchestrator) dispatchMode(state *State, issue connector.Issue) string {
-	if normalizeState(issue.State) == normalizeState(autoPromoteMergingState) && o.cfg.MergeFastPathEnabled {
-		return runpkg.RunModeMerge
-	}
-	cfg := gate.EffectivePlan(o.cfg.Plan)
-	if !cfg.Enabled {
-		return runpkg.RunModeImplement
-	}
-	switch normalizeState(issue.State) {
-	case "todo":
-		return runpkg.RunModePlan
-	case normalizeState(autoPromoteReworkState):
-		issueID := strings.TrimSpace(issue.ID)
-		if issueID != "" {
-			if _, ok := state.planRework[issueID]; ok {
-				return runpkg.RunModePlan
-			}
-		}
-		if planReviewReworkRequested(issue) {
-			return runpkg.RunModePlan
-		}
-	}
-	return runpkg.RunModeImplement
-}
-
-func (o *Orchestrator) markGlobalProjectIdle() {
-	if o.globalDispatchGate == nil {
-		return
-	}
-	o.globalDispatchGate.MarkIdle(o.cfg.Project.ID)
-}
-
-type projectStateSlotStats struct {
-	capacity  int
-	used      int
-	available int
-}
-
-type detailedProjectDispatchGate interface {
-	TryAcquireWithDecision(context.Context, scheduler.ProjectCandidate, scheduler.SlotRequest, time.Time) (
-		scheduler.Slot,
-		bool,
-		scheduler.DispatchGateDecision,
-		error,
-	)
-}
-
-func (o *Orchestrator) acquireGlobalDispatchSlot(
-	ctx context.Context,
-	issue connector.Issue,
-	workerHost string,
-	now time.Time,
-) (scheduler.Slot, bool, scheduler.DispatchGateDecision) {
-	if o.globalDispatchGate == nil {
-		return scheduler.Slot{}, true, scheduler.DispatchGateDecision{Reason: scheduler.DispatchGateReasonGranted}
-	}
-
-	req := scheduler.SlotRequest{
-		State:    issue.State,
-		Host:     workerHost,
-		Priority: o.dispatchStatePriority(issue.State),
-	}
-	var (
-		slot     scheduler.Slot
-		ok       bool
-		decision scheduler.DispatchGateDecision
-		err      error
-	)
-	if detailed, hasDecision := o.globalDispatchGate.(detailedProjectDispatchGate); hasDecision {
-		slot, ok, decision, err = detailed.TryAcquireWithDecision(ctx, o.cfg.Project, req, now)
-	} else {
-		slot, ok, err = o.globalDispatchGate.TryAcquire(ctx, o.cfg.Project, req, now)
-		if ok {
-			decision.Reason = scheduler.DispatchGateReasonGranted
-		}
-	}
-	if err != nil {
-		if o.logger != nil {
-			o.logger.Warn("global dispatch slot unavailable", "project_id", o.cfg.Project.ID, "issue_id", issue.ID, "error", err)
-		}
-		return scheduler.Slot{}, false, decision
-	}
-	return slot, ok, decision
-}
-
-func (o *Orchestrator) dispatchStatePriority(state string) int {
-	return stateDispatchRank(stateDispatchRanks(o.cfg.DispatchPriorityByState), state)
-}
-
-func (o *Orchestrator) projectStateSlotStats(issue connector.Issue, state *State) projectStateSlotStats {
-	limit := o.cfg.MaxConcurrentAgents
-	normalized := normalizeState(issue.State)
-	if stateLimit, ok := o.cfg.MaxConcurrentAgentsByState[normalized]; ok {
-		limit = stateLimit
-	}
-
-	used := 0
-	if state != nil {
-		for _, running := range state.Running {
-			if normalizeState(running.Issue.State) == normalized {
-				used++
-			}
-		}
-	}
-	available := limit - used
-	if available < 0 {
-		available = 0
-	}
-	return projectStateSlotStats{capacity: limit, used: used, available: available}
-}
-
-func (o *Orchestrator) releaseGlobalDispatchSlot(slot scheduler.Slot) {
-	if o.globalDispatchGate == nil || slot == (scheduler.Slot{}) {
-		return
-	}
-	if err := o.globalDispatchGate.Release(slot); err != nil && o.logger != nil {
-		o.logger.Warn("release global dispatch slot failed", "project_id", o.cfg.Project.ID, "error", err)
-	}
-}
-
-func (o *Orchestrator) setGlobalDispatchPreempt(slot scheduler.Slot, preempt func()) {
-	if o.globalDispatchGate == nil || slot == (scheduler.Slot{}) {
-		return
-	}
-	o.globalDispatchGate.SetPreempt(slot, preempt)
-}
-
-func (o *Orchestrator) selectorContext() selector.Context {
-	ctx := selector.Context{
-		Persona: o.cfg.SelectorPersona,
-	}
-	if identifier, ok := o.connector.(connector.InstanceIdentifier); ok {
-		ctx.InstanceLogin = identifier.InstanceLogin()
-	}
-	return ctx
-}
-
-func (o *Orchestrator) usageUpdateHandler(ctx context.Context, issueID string) runpkg.UsageUpdateHandler {
-	return func(update runpkg.UsageUpdate) error {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-
-		select {
-		case o.runUpdates <- runUpdate{issueID: issueID, usage: update}:
-			return nil
-		default:
-			return nil
-		}
-	}
-}
-
-func (o *Orchestrator) handleRunUpdate(state *State, event runUpdate) {
-	running, ok := state.Running[event.issueID]
-	if !ok {
-		return
-	}
-
-	if event.usage.SessionID != "" {
-		running.SessionID = event.usage.SessionID
-	}
-	if event.usage.TurnCount > 0 {
-		running.TurnCount = event.usage.TurnCount
-	}
-	if !event.usage.LastEventAt.IsZero() {
-		running.LastEventAt = event.usage.LastEventAt
-	}
-	if event.usage.LastEvent != "" {
-		running.LastEvent = event.usage.LastEvent
-	}
-	if event.usage.LastMessage != "" {
-		running.LastMessage = event.usage.LastMessage
-		running.LastMessageTruncation = runtimeoutput.CloneTruncation(event.usage.LastMessageTruncation)
-	}
-	if len(event.usage.RecentEvents) > 0 {
-		running.RecentEvents = cloneActivityEvents(event.usage.RecentEvents)
-	}
-	if event.usage.ProcessIdentity != "" {
-		running.ProcessIdentity = event.usage.ProcessIdentity
-	}
-	if event.usage.WorkspacePath != "" {
-		running.WorkspacePath = event.usage.WorkspacePath
-	}
-	if diffStatsPresent(event.usage.DiffStats) {
-		running.DiffStats = event.usage.DiffStats
-	}
-	running.Tokens = event.usage.Tokens
-	state.Running[event.issueID] = running
-	if event.usage.RateLimits != nil {
-		state.RateLimits = mergeRateLimits(state.RateLimits, event.usage.RateLimits)
-	}
-	if o.workAttempts != nil && running.WorkAttemptID > 0 {
-		now := event.usage.LastEventAt
-		if now.IsZero() {
-			now = time.Now()
-		}
-		heartbeat := o.runningWorkAttemptHeartbeat(state, running, now)
-		if err := o.workAttempts.RecordWorkAttemptHeartbeat(context.Background(), heartbeat); err != nil {
-			if o.logger != nil {
-				o.logger.Warn("work attempt usage heartbeat failed", "attempt_id", running.WorkAttemptID, "issue_id", event.issueID, "error", err)
-			}
-		} else {
-			o.applyWorkAttemptHeartbeatSnapshot(state, running.WorkAttemptID, heartbeat, event.usage.LastMessageTruncation)
-		}
-	}
-}
-
-func (o *Orchestrator) handleRunResult(ctx context.Context, state *State, event runpkg.Completion) {
-	running, ok := state.Running[event.IssueID]
-	if !ok {
-		return
-	}
-	o.releaseGlobalDispatchSlot(running.globalSlot)
-	o.logWorkerLifecycle(running.Issue, "worker_capacity_released",
-		"attempt", running.Attempt,
-		"worker_host", strings.TrimSpace(running.WorkerHost),
-		"reason", "run_completed",
-	)
-	running.globalSlot = scheduler.Slot{}
-	if running.cancel != nil {
-		running.cancel()
-	}
-	delete(state.Running, event.IssueID)
-
-	if workspaceIssueTerminal(running.Issue, o.cfg.TerminalStates) {
-		tokens := event.Result.Tokens
-		if tokens == (TokenTotals{}) {
-			tokens = running.Tokens
-		}
-		if diffStatsPresent(event.Result.DiffStats) {
-			running.DiffStats = event.Result.DiffStats
-		}
-		o.logWorkerLifecycle(running.Issue, "worker_"+workerOutcome(event.Err, event.Result.FinalState),
-			"attempt", running.Attempt,
-			"worker_host", strings.TrimSpace(running.WorkerHost),
-			"final_state", strings.TrimSpace(running.Issue.State),
-		)
-		o.completeTerminalRunning(context.Background(), state, event.IssueID, running, terminalCompletedAt(running.Issue, o.cfg.TerminalStates, event.CompletedAt), tokens)
-		if event.Result.RateLimits != nil {
-			state.RateLimits = mergeRateLimits(state.RateLimits, event.Result.RateLimits)
-		}
-		return
-	}
-
-	if event.Err != nil {
-		o.logWorkerLifecycle(running.Issue, "worker_"+workerOutcome(event.Err, event.Result.FinalState),
-			"attempt", running.Attempt,
-			"worker_host", strings.TrimSpace(running.WorkerHost),
-			"retry_attempt", event.RetryAttempt,
-			"retry_delay_seconds", int64(event.RetryDelay/time.Second),
-			"error", event.Err,
-		)
-		o.completeDurableWorkAttempt(ctx, state, running, event.CompletedAt, terminalStateForRun(event.Err, event.Result.FinalState), workAttemptErrorRunner, event.Err.Error(), "failed", "worker failed")
-		if mergeWorkerIssue(running.Issue) {
-			o.logMergeWorkerFailure(running.Issue, "runner_failed", event.Err)
-			o.recordMergeFailed(state, running.Issue, event.CompletedAt, "runner_failed", event.Err)
-		}
-		attempt := event.RetryAttempt
-		if attempt < 1 {
-			attempt = nextAttempt(running.Attempt)
-		}
-		if mergeWorkerIssue(running.Issue) && attempt > maxMergeWorkerRunnerFailures {
-			if o.reworkExhaustedMergeWorker(ctx, state, running, event.CompletedAt, attempt, event.Err) {
-				return
-			}
-		}
-		if o.tripInstantFailureCircuitBreaker(ctx, state, event, running, attempt) {
-			return
-		}
-		delay := event.RetryDelay
-		if delay <= 0 {
-			delay = o.retryDelay(attempt, false)
-		}
-		o.scheduleRetryAfter(
-			state,
-			running.Issue,
-			attempt,
-			event.CompletedAt,
-			delay,
-			event.Err.Error(),
-			running.WorkerHost,
-		)
-		return
-	}
-
-	if event.Request.Mode == runpkg.RunModePlan {
-		o.logWorkerLifecycle(running.Issue, "worker_"+workerOutcome(event.Err, event.Result.FinalState),
-			"attempt", running.Attempt,
-			"worker_host", strings.TrimSpace(running.WorkerHost),
-			"mode", strings.TrimSpace(event.Request.Mode),
-			"final_state", strings.TrimSpace(event.Result.FinalState),
-		)
-		o.completePlanRunning(ctx, state, event, running)
-		return
-	}
-
-	if mergeWorkerIssue(running.Issue) {
-		if o.completeLatestTerminalMergeWorkerResult(ctx, state, event, running) {
-			return
-		}
-		if state.Draining {
-			o.completeDurableWorkAttempt(ctx, state, running, event.CompletedAt, store.WorkAttemptTerminalCancelled, "draining", "worker stopped during drain", "cancelled", "worker stopped during drain")
-			o.cleanupDrainedRun(ctx, state, event.IssueID)
-			return
-		}
-		o.handleIncompleteMergeWorkerResult(ctx, state, event, running)
-		return
-	}
-
-	finalState := event.Result.FinalState
-	if finalState == "" {
-		finalState = FinalStateCompleted
-	}
-	o.logWorkerLifecycle(running.Issue, "worker_"+workerOutcome(nil, finalState),
-		"attempt", running.Attempt,
-		"worker_host", strings.TrimSpace(running.WorkerHost),
-		"mode", strings.TrimSpace(event.Request.Mode),
-		"final_state", strings.TrimSpace(finalState),
-	)
-	terminalState := terminalStateForRun(nil, finalState)
-	errorClass := ""
-	errorMessage := ""
-	if terminalState == store.WorkAttemptTerminalFailure {
-		errorClass = "runner_final_state"
-		errorMessage = finalState
-	}
-	o.completeDurableWorkAttempt(ctx, state, running, event.CompletedAt, terminalState, errorClass, errorMessage, "completed", "worker completed")
-	delete(state.InstantFailures, event.IssueID)
-
-	state.Completed[event.IssueID] = Completed{
-		Issue:       cloneIssue(running.Issue),
-		StartedAt:   running.StartedAt,
-		CompletedAt: event.CompletedAt,
-		FinalState:  finalState,
-		Tokens:      event.Result.Tokens,
-	}
-	state.TokenTotals = addTokenTotals(state.TokenTotals, event.Result.Tokens)
-	if event.Result.RateLimits != nil {
-		state.RateLimits = mergeRateLimits(state.RateLimits, event.Result.RateLimits)
-	}
-	if diffStatsPresent(event.Result.DiffStats) {
-		state.DiffStats[event.IssueID] = event.Result.DiffStats
-	}
-	if event.Result.BudgetRefusal != nil {
-		refusal := *event.Result.BudgetRefusal
-		refusal.Issue = cloneIssue(running.Issue)
-		state.BudgetRefusals[event.IssueID] = refusal
-		o.commentBudgetRefusal(ctx, event.IssueID, refusal)
-	}
-
-	if state.Draining {
-		o.cleanupDrainedRun(ctx, state, event.IssueID)
-		return
-	}
-	o.scheduleRetry(state, running.Issue, 1, event.CompletedAt, "", true, running.WorkerHost)
-}
-
-func (o *Orchestrator) commentBudgetRefusal(ctx context.Context, issueID string, refusal BudgetRefusal) {
-	if o.connector == nil {
-		return
-	}
-	body := strings.TrimSpace(refusal.Comment)
-	if body == "" {
-		body = strings.TrimSpace(refusal.Message)
-	}
-	if body == "" {
-		return
-	}
-	if err := o.connector.CreateComment(ctx, issueID, body); err != nil && o.logger != nil {
-		o.logger.Warn(
-			"budget refusal comment failed",
-			"issue_id", issueID,
-			"identifier", refusal.Issue.Identifier,
-			"code", refusal.Code,
-			"error", err,
-		)
-	}
-}
-
-func (o *Orchestrator) tripInstantFailureCircuitBreaker(
-	ctx context.Context,
-	state *State,
-	event runpkg.Completion,
-	running Running,
-	attempt int,
-) bool {
-	if state == nil || event.Err == nil || !instantFailureDuration(running, event) {
-		delete(state.InstantFailures, event.IssueID)
-		return false
-	}
-	if state.InstantFailures == nil {
-		state.InstantFailures = map[string]InstantFailure{}
-	}
-	key := instantFailureErrorKey(event.Err)
-	displayError := o.operatorText(key)
-	if displayError == "" {
-		displayError = o.operatorText(event.Err.Error())
-	}
-	failure := state.InstantFailures[event.IssueID]
-	failureKey := failure.errorKey
-	if failureKey == "" {
-		failureKey = failure.Error
-	}
-	if failureKey != key {
-		failure = InstantFailure{
-			Issue:          cloneIssue(running.Issue),
-			Error:          displayError,
-			errorKey:       key,
-			FirstFailureAt: event.CompletedAt,
-		}
-	}
-	failure.Count++
-	failure.Issue = cloneIssue(running.Issue)
-	failure.LastFailureAt = event.CompletedAt
-	state.InstantFailures[event.IssueID] = failure
-	if failure.Count < instantFailureThreshold {
-		return false
-	}
-
-	o.parkInstantFailure(ctx, state, event, running, failure, attempt)
-	return true
-}
-
-func instantFailureDuration(running Running, event runpkg.Completion) bool {
-	if !running.StartedAt.IsZero() && !event.CompletedAt.IsZero() {
-		duration := event.CompletedAt.Sub(running.StartedAt)
-		return duration >= 0 && duration < instantFailureMaxDuration
-	}
-	if event.Result.Tokens.RuntimeSeconds > 0 {
-		return event.Result.Tokens.RuntimeSeconds < instantFailureMaxDuration.Seconds()
-	}
-	return false
-}
-
-func instantFailureErrorKey(err error) string {
-	var carrier interface {
-		BackendErrorBody() string
-	}
-	if errors.As(err, &carrier) {
-		if body := strings.TrimSpace(carrier.BackendErrorBody()); body != "" {
-			return body
-		}
-	}
-	return strings.TrimSpace(err.Error())
-}
-
-func (o *Orchestrator) parkInstantFailure(
-	ctx context.Context,
-	state *State,
-	event runpkg.Completion,
-	running Running,
-	failure InstantFailure,
-	attempt int,
-) {
-	targetState := o.instantFailureParkState()
-	issue := cloneIssue(running.Issue)
-	if targetState != "" {
-		if err := o.updateIssueState(ctx, issue, targetState, event.CompletedAt, "instant_fail_circuit_breaker"); err != nil {
-			if o.logger != nil {
-				o.logger.Error(
-					"instant fail circuit breaker state transition failed",
-					"issue_id", issue.ID,
-					"identifier", issue.Identifier,
-					"target_state", targetState,
-					"error", err,
-				)
-			}
-		} else {
-			issue.State = targetState
-		}
-	}
-	if o.connector != nil {
-		if err := o.connector.CreateComment(ctx, issue.ID, instantFailureComment(issue, event.Err, failure, attempt, targetState, o.cfg.OutputTruncationMaxBytes)); err != nil && o.logger != nil {
-			o.logger.Error("instant fail circuit breaker comment failed", "issue_id", issue.ID, "identifier", issue.Identifier, "error", err)
-		}
-	}
-	if err := o.abandonClaim(ctx, issue.ID); err != nil && o.logger != nil {
-		o.logger.Error("instant fail circuit breaker claim release failed", "issue_id", issue.ID, "identifier", issue.Identifier, "error", err)
-	}
-	delete(state.Claimed, issue.ID)
-	delete(state.Retry, issue.ID)
-	delete(state.BudgetRefusals, issue.ID)
-	delete(state.PriorAttempts, issue.ID)
-	if state.Blocked == nil {
-		state.Blocked = map[string]Blocked{}
-	}
-	state.Blocked[issue.ID] = Blocked{
-		Issue:          issue,
-		Reason:         instantFailureBlockedReasonPrefix + failure.Error,
-		RecoveryReason: "fix the pinned agent model or backend configuration, then move the issue back to Todo or Rework",
-		RecoveryTarget: "Todo",
-		BlockedAt:      event.CompletedAt,
-		Source:         BlockedSourceProjectStatus,
-	}
-	recordStateEvent(state, telemetry.ActivityEvent{
-		At:      event.CompletedAt,
-		Event:   "worker_instant_fail_circuit_breaker_tripped",
-		Message: "parked " + issueLabel(issue) + " after repeated instant worker failures: " + failure.Error,
-	})
-	if o.logger != nil {
-		attrs := []any{
-			"event", "worker_instant_fail_circuit_breaker_tripped",
-			"issue_id", issue.ID,
-			"issue_identifier", issue.Identifier,
-			"attempt", attempt,
-			"instant_failures", failure.Count,
-			"target_state", targetState,
-			"error", event.Err,
-		}
-		if body := o.operatorText(instantFailureErrorKey(event.Err)); body != "" {
-			attrs = append(attrs, "backend_error_body", body)
-		}
-		var carrier interface {
-			BackendErrorMessage() string
-		}
-		if errors.As(event.Err, &carrier) {
-			if message := o.operatorText(carrier.BackendErrorMessage()); message != "" {
-				attrs = append(attrs, "backend_error_message", message)
-			}
-		}
-		o.logger.Error("worker instant fail circuit breaker tripped", attrs...)
-	}
-}
-
-func (o *Orchestrator) instantFailureParkState() string {
-	return blockedStatusState
-}
-
-func instantFailureComment(issue connector.Issue, err error, failure InstantFailure, attempt int, targetState string, maxBytes int) string {
-	var b strings.Builder
-	b.WriteString("Detent stopped retrying this worker after ")
-	b.WriteString(strconv.Itoa(failure.Count))
-	b.WriteString(" consecutive instant failures with the same backend error.")
-	if targetState = strings.TrimSpace(targetState); targetState != "" {
-		b.WriteString("\n\nIssue parked in `")
-		b.WriteString(targetState)
-		b.WriteString("`.")
-	}
-	b.WriteString("\n\n- issue: ")
-	b.WriteString(issueLabel(issue))
-	b.WriteString("\n- latest_attempt: ")
-	b.WriteString(strconv.Itoa(attempt))
-	b.WriteString("\n- failure_window_seconds: ")
-	b.WriteString(strconv.FormatInt(int64(instantFailureMaxDuration/time.Second), 10))
-	b.WriteString("\n- error:\n\n```text\n")
-	errorText := runtimeoutput.Truncate(strings.TrimSpace(err.Error()), maxBytes).Value
-	b.WriteString(errorText)
-	b.WriteString("\n```")
-	if body := runtimeoutput.Truncate(instantFailureErrorKey(err), maxBytes).Value; body != "" && body != errorText {
-		b.WriteString("\n\n- backend_error_body:\n\n```json\n")
-		b.WriteString(body)
-		b.WriteString("\n```")
-	}
-	b.WriteString("\n\nFix the pinned agent model or backend configuration, then move the issue back to Todo or Rework.")
-	return b.String()
-}
-
-func (o *Orchestrator) operatorText(value string) string {
-	return runtimeoutput.Truncate(strings.TrimSpace(value), o.cfg.OutputTruncationMaxBytes).Value
-}
-
 func (o *Orchestrator) cleanupDrainedRun(ctx context.Context, state *State, issueID string) {
 	if err := o.abandonClaim(ctx, issueID); err != nil && o.logger != nil {
 		o.logger.Warn("abandon completed drain claim failed", "issue_id", issueID, "error", err)
@@ -2338,837 +515,4 @@ func (o *Orchestrator) cleanupDrainedRun(ctx context.Context, state *State, issu
 	delete(state.BudgetRefusals, issueID)
 	delete(state.PriorAttempts, issueID)
 	delete(state.InstantFailures, issueID)
-}
-
-func (o *Orchestrator) completeLatestTerminalMergeWorkerResult(
-	ctx context.Context,
-	state *State,
-	event runpkg.Completion,
-	running Running,
-) bool {
-	issueID := strings.TrimSpace(event.IssueID)
-	if issueID == "" || o.connector == nil {
-		return false
-	}
-	issues, err := o.connector.FetchIssueStatesByIDs(ctx, []string{issueID})
-	if err != nil {
-		if o.logger != nil {
-			o.logger.Warn(
-				"merge_worker_terminal_state_refresh_failed",
-				"issue_id", issueID,
-				"identifier", running.Issue.Identifier,
-				"error", err,
-			)
-		}
-		return false
-	}
-	for _, issue := range issues {
-		if strings.TrimSpace(issue.ID) != issueID {
-			continue
-		}
-		issue = mergeIssueTrackerFields(running.Issue, issue)
-		if !workspaceIssueTerminal(issue, o.cfg.TerminalStates) {
-			return o.completeProgrammaticMergeWorkerResult(ctx, state, event, running, issue)
-		}
-		tokens := event.Result.Tokens
-		if tokens == (TokenTotals{}) {
-			tokens = running.Tokens
-		}
-		if diffStatsPresent(event.Result.DiffStats) {
-			running.DiffStats = event.Result.DiffStats
-		}
-		running.Issue = issue
-		o.completeTerminalRunning(ctx, state, issueID, running, terminalCompletedAt(issue, o.cfg.TerminalStates, event.CompletedAt), tokens)
-		if event.Result.RateLimits != nil {
-			state.RateLimits = mergeRateLimits(state.RateLimits, event.Result.RateLimits)
-		}
-		return true
-	}
-	return false
-}
-
-func (o *Orchestrator) completeProgrammaticMergeWorkerResult(
-	ctx context.Context,
-	state *State,
-	event runpkg.Completion,
-	running Running,
-	issue connector.Issue,
-) bool {
-	if state != nil && state.Draining {
-		return false
-	}
-	if !mergeWorkerTurnSucceeded(event) {
-		return false
-	}
-	hydrator, ok := o.connector.(connector.PullRequestHydrator)
-	if !ok {
-		return false
-	}
-	refreshedIssue, err := hydrator.HydratePullRequest(ctx, issue)
-	if err != nil {
-		running.Issue = issue
-		o.failProgrammaticMergeWorkerResult(ctx, state, event, running, "programmatic_merge_pr_refresh_failed", err)
-		return true
-	}
-	issue = refreshedIssue
-	if !mergeWorkerProgrammaticMergeReady(issue) {
-		if mergeFastPathCleanResult(event) && mergeWorkerProgrammaticMergeWaiting(issue) {
-			o.waitForMergeWorkerCurrentHeadCI(ctx, state, event, running, issue)
-			return true
-		}
-		return false
-	}
-	merger, ok := o.connector.(connector.PullRequestMerger)
-	if !ok {
-		return false
-	}
-	issueID := strings.TrimSpace(event.IssueID)
-	repository := pullRequestRepository(issue)
-	number := pullRequestNumber(issue)
-	headSHA := strings.TrimSpace(issue.PullRequest.HeadSHA)
-	if err := merger.MergePullRequest(ctx, repository, number, headSHA); err != nil {
-		running.Issue = issue
-		o.failProgrammaticMergeWorkerResult(ctx, state, event, running, "programmatic_merge_failed", err)
-		return true
-	}
-
-	targetState := doneStateName(o.cfg.TerminalStates)
-	mergedIssue := cloneIssue(issue)
-	if mergedIssue.PullRequest != nil {
-		mergedIssue.PullRequest.State = "MERGED"
-		activityAt := event.CompletedAt.UTC()
-		mergedIssue.PullRequest.ActivityAt = &activityAt
-	}
-	if err := o.updateIssueStateByID(ctx, issueID, mergedIssue, targetState, event.CompletedAt, "merge_worker_programmatic_merge"); err != nil {
-		running.Issue = mergedIssue
-		o.failProgrammaticMergeWorkerResult(ctx, state, event, running, "programmatic_merge_state_update_failed", err)
-		return true
-	}
-
-	updatedAt := event.CompletedAt.UTC()
-	mergedIssue.State = targetState
-	mergedIssue.UpdatedAt = &updatedAt
-	mergedIssue.StageUpdatedAt = &updatedAt
-	mergeTimingIssue := running.Issue
-	running.Issue = mergedIssue
-	tokens := event.Result.Tokens
-	if tokens == (TokenTotals{}) {
-		tokens = running.Tokens
-	}
-	if diffStatsPresent(event.Result.DiffStats) {
-		running.DiffStats = event.Result.DiffStats
-	}
-	if o.logger != nil {
-		o.logger.Info("merge_worker_programmatic_merge", mergeWorkerLogAttrs(mergedIssue, "target_state", targetState)...)
-	}
-	recordStateEvent(state, telemetry.ActivityEvent{
-		At:      event.CompletedAt,
-		Event:   "merge_worker_programmatic_merge",
-		Message: "programmatically merged " + issueLabel(mergedIssue) + " and moved it to " + targetState,
-	})
-	o.completeTerminalRunning(ctx, state, issueID, running, terminalCompletedAt(mergedIssue, o.cfg.TerminalStates, event.CompletedAt), tokens)
-	mergeTiming := o.recordMergeCompleted(state, mergeTimingIssue, event.CompletedAt, targetState)
-	if completed, ok := state.Completed[issueID]; ok {
-		completed.MergeTiming = mergeTiming
-		state.Completed[issueID] = completed
-	}
-	o.logMergeWorkerSuccess(mergeTimingIssue, targetState)
-	if event.Result.RateLimits != nil {
-		state.RateLimits = mergeRateLimits(state.RateLimits, event.Result.RateLimits)
-	}
-	return true
-}
-
-func (o *Orchestrator) waitForMergeWorkerCurrentHeadCI(
-	ctx context.Context,
-	state *State,
-	event runpkg.Completion,
-	running Running,
-	issue connector.Issue,
-) {
-	running.Issue = issue
-	o.completeDurableWorkAttempt(ctx, state, running, event.CompletedAt, store.WorkAttemptTerminalSuccess, "", "", "waiting", "merge fast-path waiting for current-head CI")
-	attempt := running.Attempt
-	if attempt < 1 {
-		attempt = 1
-	}
-	o.scheduleRetry(state, issue, attempt, event.CompletedAt, "waiting for current-head CI", true, running.WorkerHost)
-	if o.logger != nil {
-		o.logger.Info("merge_worker_waiting_current_head_ci", mergeWorkerLogAttrs(issue, "attempt", attempt)...)
-	}
-	recordStateEvent(state, telemetry.ActivityEvent{
-		At:      event.CompletedAt,
-		Event:   "merge_worker_waiting_current_head_ci",
-		Message: "merge worker is waiting for current-head CI for " + issueLabel(issue),
-	})
-}
-
-func (o *Orchestrator) failProgrammaticMergeWorkerResult(
-	ctx context.Context,
-	state *State,
-	event runpkg.Completion,
-	running Running,
-	reason string,
-	err error,
-) {
-	o.completeDurableWorkAttempt(ctx, state, running, event.CompletedAt, store.WorkAttemptTerminalFailure, reason, errorString(err), "merging", "programmatic merge failed")
-	o.logMergeWorkerFailure(running.Issue, reason, err)
-	o.recordMergeFailed(state, running.Issue, event.CompletedAt, reason, err)
-	attempt := nextAttempt(running.Attempt)
-	if attempt > maxMergeWorkerRunnerFailures {
-		if o.reworkExhaustedMergeWorker(ctx, state, running, event.CompletedAt, attempt, err) {
-			return
-		}
-	}
-	o.scheduleRetry(
-		state,
-		running.Issue,
-		attempt,
-		event.CompletedAt,
-		errorString(err),
-		false,
-		running.WorkerHost,
-	)
-	recordStateEvent(state, telemetry.ActivityEvent{
-		At:      event.CompletedAt,
-		Event:   "merge_worker_programmatic_merge_failed",
-		Message: "programmatic merge failed for " + issueLabel(running.Issue) + ": " + errorString(err),
-	})
-}
-
-func mergeWorkerTurnSucceeded(event runpkg.Completion) bool {
-	return event.Err == nil && !strings.EqualFold(strings.TrimSpace(event.Result.FinalState), runpkg.FinalStateFailed)
-}
-
-func mergeFastPathCleanResult(event runpkg.Completion) bool {
-	return event.Request.Mode == runpkg.RunModeMerge &&
-		strings.TrimSpace(event.Result.Output) == runpkg.RunOutputMergeFastPathClean
-}
-
-func mergeWorkerProgrammaticMergeReady(issue connector.Issue) bool {
-	if strings.TrimSpace(issue.ID) == "" || issue.PullRequest == nil {
-		return false
-	}
-	pullRequest := issue.PullRequest
-	if pullRequestHydrationBlocksProgress(pullRequest) {
-		return false
-	}
-	if normalizePullRequestState(pullRequest.State) != "open" || pullRequest.Draft {
-		return false
-	}
-	if strings.ToLower(strings.TrimSpace(pullRequest.MergeableState)) != "clean" {
-		return false
-	}
-	if !mergeWorkerCIGreen(pullRequest.CIStatus) {
-		return false
-	}
-	return pullRequestRepository(issue) != "" &&
-		pullRequestNumber(issue) > 0 &&
-		strings.TrimSpace(pullRequest.HeadSHA) != ""
-}
-
-func mergeWorkerProgrammaticMergeWaiting(issue connector.Issue) bool {
-	if strings.TrimSpace(issue.ID) == "" || issue.PullRequest == nil {
-		return false
-	}
-	pullRequest := issue.PullRequest
-	if pullRequestHydrationBlocksProgress(pullRequest) {
-		return false
-	}
-	if normalizePullRequestState(pullRequest.State) != "open" || pullRequest.Draft {
-		return false
-	}
-	if mergeable := strings.ToLower(strings.TrimSpace(pullRequest.MergeableState)); mergeable != "" && mergeable != "clean" && mergeable != "unknown" {
-		return false
-	}
-	if pullRequestRepository(issue) == "" || pullRequestNumber(issue) <= 0 || strings.TrimSpace(pullRequest.HeadSHA) == "" {
-		return false
-	}
-	switch strings.ToLower(strings.TrimSpace(pullRequest.CIStatus)) {
-	case "", "pending", "running", "queued", "in_progress", "waiting":
-		return true
-	default:
-		return false
-	}
-}
-
-func mergeWorkerCIGreen(status string) bool {
-	switch strings.ToLower(strings.TrimSpace(status)) {
-	case "success", "green", "pass", "passed":
-		return true
-	default:
-		return false
-	}
-}
-
-func (o *Orchestrator) handleIncompleteMergeWorkerResult(
-	ctx context.Context,
-	state *State,
-	event runpkg.Completion,
-	running Running,
-) {
-	err := errors.New(mergeWorkerTerminalStateMissing)
-	o.completeDurableWorkAttempt(ctx, state, running, event.CompletedAt, store.WorkAttemptTerminalFailure, workAttemptErrorMergeIncomplete, err.Error(), "merging", "merge worker completed without terminal state")
-	o.logMergeWorkerFailure(running.Issue, "terminal_state_missing", err)
-	attempt := nextAttempt(running.Attempt)
-	if attempt > maxMergeWorkerRunnerFailures {
-		if o.reworkExhaustedMergeWorker(ctx, state, running, event.CompletedAt, attempt, err) {
-			return
-		}
-	}
-	o.scheduleRetry(
-		state,
-		running.Issue,
-		attempt,
-		event.CompletedAt,
-		err.Error(),
-		false,
-		running.WorkerHost,
-	)
-	recordStateEvent(state, telemetry.ActivityEvent{
-		At:      event.CompletedAt,
-		Event:   "merge_worker_terminal_state_missing",
-		Message: "merge worker completed without terminal state for " + issueLabel(running.Issue),
-	})
-}
-
-func (o *Orchestrator) reworkExhaustedMergeWorker(
-	ctx context.Context,
-	state *State,
-	running Running,
-	completedAt time.Time,
-	attempt int,
-	err error,
-) bool {
-	issueID := strings.TrimSpace(running.Issue.ID)
-	if issueID == "" || o.connector == nil {
-		return false
-	}
-	if err := o.updateIssueStateByID(ctx, issueID, running.Issue, autoPromoteReworkState, completedAt, "merge_worker_retry_exhausted"); err != nil {
-		if o.logger != nil {
-			o.logger.Warn(
-				"merge_worker_rework_failed",
-				"issue_id", issueID,
-				"identifier", running.Issue.Identifier,
-				"reason", "runner_failed_retry_exhausted",
-				"target_state", autoPromoteReworkState,
-				"error", err,
-			)
-		}
-		return false
-	}
-	if comment := mergeWorkerRetryExhaustedComment(running.Issue, attempt, err); strings.TrimSpace(comment) != "" {
-		if err := o.connector.CreateComment(ctx, issueID, comment); err != nil && o.logger != nil {
-			o.logger.Warn(
-				"merge_worker_rework_comment_failed",
-				"issue_id", issueID,
-				"identifier", running.Issue.Identifier,
-				"reason", "runner_failed_retry_exhausted",
-				"error", err,
-			)
-		}
-	}
-	if err := o.abandonClaim(ctx, issueID); err != nil && o.logger != nil {
-		o.logger.Warn("abandon exhausted merge worker claim failed", "issue_id", issueID, "error", err)
-	}
-	delete(state.Claimed, issueID)
-	delete(state.Retry, issueID)
-	delete(state.BudgetRefusals, issueID)
-	delete(state.PriorAttempts, issueID)
-	delete(state.Completed, issueID)
-	recordStateEvent(state, telemetry.ActivityEvent{
-		At:      completedAt,
-		Event:   "merge_worker_retry_exhausted",
-		Message: "merge worker retries exhausted for " + issueLabel(running.Issue) + ": " + errorString(err),
-	})
-	return true
-}
-
-func mergeWorkerRetryExhaustedComment(issue connector.Issue, attempt int, err error) string {
-	var b strings.Builder
-	b.WriteString("Merge worker retries were exhausted; routed this issue from Merging to Rework.")
-	b.WriteString("\n\n- reason: runner_failed_retry_exhausted")
-	if attempt > 0 {
-		b.WriteString("\n- attempt: ")
-		b.WriteString(strconv.Itoa(attempt))
-	}
-	if errText := errorString(err); errText != "" {
-		b.WriteString("\n- error: ")
-		b.WriteString(errText)
-	}
-	if issue.PullRequest != nil {
-		if url := strings.TrimSpace(issue.PullRequest.URL); url != "" {
-			b.WriteString("\n- pull request: ")
-			b.WriteString(url)
-		}
-		if mergeableState := strings.ToLower(strings.TrimSpace(issue.PullRequest.MergeableState)); mergeableState != "" {
-			b.WriteString("\n- mergeable_state: ")
-			b.WriteString(mergeableState)
-		}
-		if ciStatus := strings.TrimSpace(issue.PullRequest.CIStatus); ciStatus != "" {
-			b.WriteString("\n- ci_status: ")
-			b.WriteString(ciStatus)
-		}
-	}
-	return b.String()
-}
-
-func errorString(err error) string {
-	if err == nil {
-		return ""
-	}
-	return strings.TrimSpace(err.Error())
-}
-
-func (o *Orchestrator) completePlanRunning(
-	ctx context.Context,
-	state *State,
-	event runpkg.Completion,
-	running Running,
-) {
-	cfg := gate.EffectivePlan(o.cfg.Plan)
-	issueID := strings.TrimSpace(event.IssueID)
-	issue := cloneIssue(running.Issue)
-	body := planArtifactComment(issue, event.Result.Output)
-	if err := o.connector.CreateComment(ctx, issueID, body); err != nil {
-		o.completeDurableWorkAttempt(ctx, state, running, event.CompletedAt, store.WorkAttemptTerminalFailure, "plan_comment_failed", err.Error(), "reviewing", "plan comment failed")
-		o.scheduleRetry(state, issue, nextAttempt(running.Attempt), event.CompletedAt, "plan comment failed: "+err.Error(), false, running.WorkerHost)
-		return
-	}
-	if err := o.updateIssueStateByID(ctx, issueID, issue, cfg.Stop, event.CompletedAt, "plan_artifact_created"); err != nil {
-		o.completeDurableWorkAttempt(ctx, state, running, event.CompletedAt, store.WorkAttemptTerminalFailure, "plan_transition_failed", err.Error(), "reviewing", "plan review transition failed")
-		o.scheduleRetry(state, issue, nextAttempt(running.Attempt), event.CompletedAt, "plan review transition failed: "+err.Error(), false, running.WorkerHost)
-		return
-	}
-	o.completeDurableWorkAttempt(ctx, state, running, event.CompletedAt, store.WorkAttemptTerminalSuccess, "", "", "completed", "plan review created")
-	if err := o.abandonClaim(ctx, issueID); err != nil && o.logger != nil {
-		o.logger.Warn("abandon completed plan claim failed", "issue_id", issueID, "error", err)
-	}
-	delete(state.planRework, issueID)
-	issue.State = cfg.Stop
-	state.Completed[issueID] = Completed{
-		Issue:       issue,
-		StartedAt:   running.StartedAt,
-		CompletedAt: event.CompletedAt,
-		FinalState:  cfg.Stop,
-		Tokens:      event.Result.Tokens,
-	}
-	state.TokenTotals = addTokenTotals(state.TokenTotals, event.Result.Tokens)
-	if event.Result.RateLimits != nil {
-		state.RateLimits = mergeRateLimits(state.RateLimits, event.Result.RateLimits)
-	}
-	if diffStatsPresent(event.Result.DiffStats) {
-		state.DiffStats[issueID] = event.Result.DiffStats
-	}
-	delete(state.Claimed, issueID)
-	delete(state.Retry, issueID)
-	delete(state.BudgetRefusals, issueID)
-	delete(state.PriorAttempts, issueID)
-	recordStateEvent(state, telemetry.ActivityEvent{
-		At:      event.CompletedAt,
-		Event:   "plan_review_created",
-		Message: "created plan artifact for " + issueLabel(issue) + " and moved to " + cfg.Stop,
-	})
-}
-
-func (o *Orchestrator) scheduleRetry(
-	state *State,
-	issue connector.Issue,
-	attempt int,
-	now time.Time,
-	err string,
-	continuation bool,
-	workerHost string,
-) {
-	o.dispatchPlanner().scheduleRetry(state, issue, attempt, now, err, continuation, workerHost)
-}
-
-func (o *Orchestrator) scheduleRetryAfter(
-	state *State,
-	issue connector.Issue,
-	attempt int,
-	now time.Time,
-	delay time.Duration,
-	err string,
-	workerHost string,
-) {
-	o.dispatchPlanner().scheduleRetryAfter(state, issue, attempt, now, delay, err, workerHost)
-}
-
-func (o *Orchestrator) retryDelay(attempt int, continuation bool) time.Duration {
-	return o.dispatchPlanner().retryDelay(attempt, continuation)
-}
-
-func (o *Orchestrator) releaseClaim(state *State, issueID string) {
-	o.cancelRunning(state, issueID)
-	delete(state.Running, issueID)
-	delete(state.Claimed, issueID)
-	delete(state.Retry, issueID)
-	delete(state.BudgetRefusals, issueID)
-	delete(state.PriorAttempts, issueID)
-}
-
-func (o *Orchestrator) completeTerminalRunning(
-	ctx context.Context,
-	state *State,
-	issueID string,
-	running Running,
-	completedAt time.Time,
-	tokens TokenTotals,
-) {
-	o.completeDurableWorkAttempt(ctx, state, running, completedAt, store.WorkAttemptTerminalSuccess, "", "", "completed", "worker reached terminal state")
-	o.releaseGlobalDispatchSlot(running.globalSlot)
-	if running.cancel != nil {
-		running.cancel()
-	}
-	delete(state.Running, issueID)
-	delete(state.Claimed, issueID)
-	delete(state.Retry, issueID)
-	delete(state.BudgetRefusals, issueID)
-	delete(state.PriorAttempts, issueID)
-	delete(state.InstantFailures, issueID)
-	if err := o.abandonClaim(ctx, issueID); err != nil {
-		recordStateEvent(state, telemetry.ActivityEvent{
-			At:      cleanupEventAt(completedAt),
-			Event:   "claim_release_failed",
-			Message: fmt.Sprintf("claim lease release failed for %s: %v", issueLabel(running.Issue), err),
-		})
-	}
-	issue := o.ensureClosedCompletedRunningIssueDone(ctx, issueID, running.Issue, completedAt)
-	finalState := strings.TrimSpace(issue.State)
-	if finalState == "" {
-		finalState = FinalStateCompleted
-	}
-	mergeTiming := MergeTiming{}
-	if mergeWorkerIssue(running.Issue) {
-		mergeTiming = o.recordMergeCompleted(state, running.Issue, completedAt, finalState)
-	}
-	state.Completed[issueID] = Completed{
-		Issue:       cloneIssue(issue),
-		StartedAt:   running.StartedAt,
-		CompletedAt: completedAt,
-		FinalState:  finalState,
-		Tokens:      tokens,
-		MergeTiming: mergeTiming,
-	}
-	state.TokenTotals = addTokenTotals(state.TokenTotals, tokens)
-	if diffStatsPresent(running.DiffStats) {
-		state.DiffStats[issueID] = running.DiffStats
-	}
-	if mergeWorkerIssue(running.Issue) {
-		o.logMergeWorkerSuccess(running.Issue, finalState)
-	}
-	o.reapWorkspace(ctx, state, issue, workspaceReapReason(issue, o.cfg.TerminalStates), completedAt)
-}
-
-func (o *Orchestrator) ensureClosedCompletedRunningIssueDone(ctx context.Context, issueID string, issue connector.Issue, now time.Time) connector.Issue {
-	if !issue.Closed || !closedReasonCompleted(issue.ClosedReason) {
-		return issue
-	}
-	targetState := doneStateName(o.cfg.TerminalStates)
-	if strings.TrimSpace(targetState) == "" {
-		return issue
-	}
-	if err := o.updateIssueStateByID(ctx, issueID, issue, targetState, now, "closed_completed_running_done"); err != nil {
-		if o.logger != nil {
-			o.logger.Warn("mark closed completed running issue done failed", "issue_id", issueID, "identifier", issue.Identifier, "from_state", issue.State, "target_state", targetState, "error", err)
-		}
-		return issue
-	}
-	if o.logger != nil {
-		o.logger.Info("marked closed completed running issue done", "issue_id", issueID, "identifier", issue.Identifier, "from_state", issue.State, "target_state", targetState)
-	}
-	issue.State = targetState
-	return issue
-}
-
-func terminalCompletedAt(issue connector.Issue, terminalStates []string, fallback time.Time) time.Time {
-	if stateIn(issue.State, terminalStates) && issue.StageUpdatedAt != nil && !issue.StageUpdatedAt.IsZero() {
-		return *issue.StageUpdatedAt
-	}
-	if issue.UpdatedAt != nil && !issue.UpdatedAt.IsZero() {
-		return *issue.UpdatedAt
-	}
-	if issue.StageUpdatedAt != nil && !issue.StageUpdatedAt.IsZero() {
-		return *issue.StageUpdatedAt
-	}
-	if !fallback.IsZero() {
-		return fallback
-	}
-	return time.Now().UTC()
-}
-
-func (o *Orchestrator) cancelRunning(state *State, issueID string) {
-	running, ok := state.Running[issueID]
-	if !ok {
-		return
-	}
-	o.releaseGlobalDispatchSlot(running.globalSlot)
-	running.globalSlot = scheduler.Slot{}
-	state.Running[issueID] = running
-	cancelRunning(state, issueID)
-}
-
-func cancelRunning(state *State, issueID string) {
-	running, ok := state.Running[issueID]
-	if !ok || running.cancel == nil {
-		return
-	}
-	running.cancel()
-	running.cancel = nil
-	state.Running[issueID] = running
-}
-
-func (o *Orchestrator) releaseRunningSlots(state *State) {
-	for issueID, running := range state.Running {
-		o.releaseGlobalDispatchSlot(running.globalSlot)
-		running.globalSlot = scheduler.Slot{}
-		state.Running[issueID] = running
-	}
-}
-
-func normalizeConfig(cfg Config) Config {
-	if cfg.PollInterval <= 0 {
-		cfg.PollInterval = defaultPollInterval
-	}
-	if cfg.MaxConcurrentAgents <= 0 {
-		cfg.MaxConcurrentAgents = defaultMaxConcurrentAgents
-	}
-	if cfg.MaxRetryBackoff <= 0 {
-		cfg.MaxRetryBackoff = defaultMaxRetryBackoff
-	}
-	if cfg.ContinuationRetryDelay < 0 {
-		cfg.ContinuationRetryDelay = 0
-	}
-	if cfg.ContinuationRetryDelay == 0 {
-		cfg.ContinuationRetryDelay = defaultContinuationRetry
-	}
-	if cfg.FailureRetryBaseDelay <= 0 {
-		cfg.FailureRetryBaseDelay = defaultFailureRetryBaseDelay
-	}
-	if cfg.GitHubGraphQLWarnRemaining <= 0 {
-		cfg.GitHubGraphQLWarnRemaining = defaultGitHubGraphQLWarnRemaining
-	}
-	if cfg.GitHubGraphQLMinReserve <= 0 {
-		cfg.GitHubGraphQLMinReserve = defaultGitHubGraphQLMinReserve
-	}
-	if cfg.GitHubRESTMinReserve <= 0 {
-		cfg.GitHubRESTMinReserve = defaultGitHubRESTMinReserve
-	}
-	if len(cfg.ActiveStates) == 0 {
-		cfg.ActiveStates = []string{"Todo", "In Progress"}
-	}
-	if len(cfg.TerminalStates) == 0 {
-		cfg.TerminalStates = []string{"Done", "Cancelled", "Canceled", "Closed"}
-	}
-	if cfg.WorkspaceCleanupIdleTTL <= 0 {
-		cfg.WorkspaceCleanupIdleTTL = defaultWorkspaceCleanupIdleTTL
-	}
-	if cfg.WorkspaceCleanupSweepInterval <= 0 {
-		cfg.WorkspaceCleanupSweepInterval = defaultWorkspaceCleanupSweep
-	}
-
-	cfg.ActiveStates = normalizedStates(cfg.ActiveStates)
-	cfg.ObservedStates = normalizedStates(cfg.ObservedStates)
-	cfg.TerminalStates = normalizedStates(cfg.TerminalStates)
-	cfg.MaxConcurrentAgentsByState = cloneStateLimits(cfg.MaxConcurrentAgentsByState)
-	cfg.DispatchPriorityByState = normalizedStates(cfg.DispatchPriorityByState)
-	cfg.DispatchPriorityByLabel = normalizeLabels(cfg.DispatchPriorityByLabel)
-	cfg.Claiming = normalizeClaimingConfig(cfg.Claiming)
-	cfg.AutoPromote = normalizeAutoPromoteConfig(cfg.AutoPromote)
-	cfg.Plan = gate.EffectivePlan(cfg.Plan)
-	cfg.DependencyAutoUnblock = normalizeDependencyAutoUnblockConfig(cfg.DependencyAutoUnblock)
-	cfg.BlockerAutoPromote = normalizeBlockerAutoPromoteConfig(cfg.BlockerAutoPromote, cfg.ActiveStates, cfg.DependencyAutoUnblock)
-	cfg.Authorization.Normalize()
-	cfg.SelectorContext.InstanceLogin = strings.TrimSpace(cfg.SelectorContext.InstanceLogin)
-	cfg.SelectorContext.Persona = strings.TrimSpace(cfg.SelectorContext.Persona)
-	cfg.WorkerHosts = normalizeWorkerHosts(cfg.WorkerHosts)
-	cfg.SelectorPersona = strings.TrimSpace(cfg.SelectorPersona)
-	if cfg.MaxConcurrentAgentsPerHost < 0 {
-		cfg.MaxConcurrentAgentsPerHost = 0
-	}
-	if cfg.OutputTruncationMaxBytes < 0 {
-		cfg.OutputTruncationMaxBytes = 0
-	}
-
-	return cfg
-}
-
-func normalizeClaimingConfig(cfg ClaimingConfig) ClaimingConfig {
-	cfg.OwnershipMode = strings.ToLower(strings.TrimSpace(cfg.OwnershipMode))
-	if cfg.OwnershipMode == "" {
-		cfg.OwnershipMode = workflowconfig.IdentityOwnershipAssignee
-	}
-	cfg.Owner = strings.TrimSpace(cfg.Owner)
-	cfg.AssigneeLogin = strings.TrimSpace(cfg.AssigneeLogin)
-	cfg.OwnerField = strings.TrimSpace(cfg.OwnerField)
-	cfg.LeaseField = strings.TrimSpace(cfg.LeaseField)
-	if cfg.LeaseTTL < 0 {
-		cfg.LeaseTTL = 0
-	}
-	if cfg.HeartbeatInterval < 0 {
-		cfg.HeartbeatInterval = 0
-	}
-	return cfg
-}
-
-func durationFromMillis(ms int) time.Duration {
-	if ms <= 0 {
-		return 0
-	}
-	return time.Duration(ms) * time.Millisecond
-}
-
-func durationFromSeconds(seconds int) time.Duration {
-	if seconds <= 0 {
-		return 0
-	}
-	return time.Duration(seconds) * time.Second
-}
-
-func positiveIntValue(value *int) int {
-	if value == nil || *value <= 0 {
-		return 0
-	}
-	return *value
-}
-
-func cloneStateLimits(limits map[string]int) map[string]int {
-	cloned := make(map[string]int, len(limits))
-	for state, limit := range limits {
-		if limit > 0 {
-			cloned[normalizeState(state)] = limit
-		}
-	}
-	return cloned
-}
-
-func cloneIssues(issues []connector.Issue) []connector.Issue {
-	cloned := make([]connector.Issue, len(issues))
-	for i, issue := range issues {
-		cloned[i] = cloneIssue(issue)
-	}
-	return cloned
-}
-
-func validCandidate(issue connector.Issue) bool {
-	return issue.ID != "" &&
-		issue.Identifier != "" &&
-		issue.Title != "" &&
-		issue.State != "" &&
-		!issue.Closed &&
-		issue.AssignedToWorker
-}
-
-func duplicatePullRequestWork(issue connector.Issue) bool {
-	if issue.PullRequest == nil {
-		return false
-	}
-	switch normalizePullRequestState(issue.PullRequest.State) {
-	case "merged":
-		return !staleMergedPullRequestHasFailedCIEvidence(issue.PullRequest, staleMergedPullRequestSummaryFromIssue(issue))
-	case "open":
-		return normalizeState(issue.State) == "todo"
-	default:
-		return false
-	}
-}
-
-func mergedPullRequestReconciliationPending(issue connector.Issue, cfg Config) bool {
-	if issue.PullRequest == nil || normalizePullRequestState(issue.PullRequest.State) != "merged" {
-		return false
-	}
-	summary := staleMergedPullRequestSummaryFromIssue(issue)
-	decision := staleMergedPullRequestDecision(issue, summary)
-	targetState := staleMergedPullRequestTargetState(decision, cfg.AutoPromote, cfg.TerminalStates)
-	return targetState != "" && normalizeState(targetState) != normalizeState(issue.State)
-}
-
-func continuationDispatch(issue connector.Issue) bool {
-	state := normalizeState(issue.State)
-	return state != "" && state != "todo"
-}
-
-func continuationDelay(index int) time.Duration {
-	if index <= 0 {
-		return 0
-	}
-	return continuationDispatchBackoff
-}
-
-func waitForDispatchBackoff(ctx context.Context, delay time.Duration) bool {
-	if delay <= 0 {
-		return true
-	}
-
-	timer := time.NewTimer(delay)
-	defer timer.Stop()
-
-	select {
-	case <-ctx.Done():
-		return false
-	case <-timer.C:
-		return true
-	}
-}
-
-func normalizePullRequestState(state string) string {
-	return strings.ToLower(strings.TrimSpace(state))
-}
-
-func todoBlockedByNonTerminal(issue connector.Issue, terminalStates []string) bool {
-	if normalizeState(issue.State) != "todo" {
-		return false
-	}
-
-	for _, blocker := range issue.BlockedBy {
-		if strings.TrimSpace(blocker.State) == "" {
-			continue
-		}
-		if !stateIn(blocker.State, terminalStates) {
-			return true
-		}
-	}
-	return false
-}
-
-func availableSlots(state *State) int {
-	available := state.MaxConcurrentAgents - len(state.Running)
-	if available < 0 {
-		return 0
-	}
-	return available
-}
-
-func nextAttempt(attempt int) int {
-	if attempt < 1 {
-		return 1
-	}
-	return attempt + 1
-}
-
-func stateIn(state string, states []string) bool {
-	normalized := normalizeState(state)
-	return slices.Contains(states, normalized)
-}
-
-func normalizedStates(states []string) []string {
-	normalized := make([]string, 0, len(states))
-	seen := make(map[string]struct{}, len(states))
-	for _, state := range states {
-		state = normalizeState(state)
-		if state == "" {
-			continue
-		}
-		if _, ok := seen[state]; ok {
-			continue
-		}
-		seen[state] = struct{}{}
-		normalized = append(normalized, state)
-	}
-	return normalized
-}
-
-func normalizeState(state string) string {
-	return strings.ToLower(strings.TrimSpace(state))
 }
